==> scripts/1.1_Field-data-organization.Rmd <==
---
title: "Extract Field Data"
author: "Anna Talucci"
date: "2/27/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Overview

This script is reliant on
* `Rproj` and therefore the `wd()`does not have to be set
* Relative path names "../folder-name/file-name.ext"
* This script is reliant on the following folders 
    +"../scripts/" --this houses all scripts written in RMarkdown
    +"../data/shapfiles/" --this houses shape files
    +"../figures/

**Data organization** 
Involve converting to plot level varaibles and some cleaning. Create plot_dis for distance on the transect that removes alpha 'a' and 'b' from -25 plots on BP transect.

Flight Code      |	Field Site	  | Transect 	| UAV Flight	 
-----------------|----------------|-----------|----------------
alnus_tr1_f07    |  Alnus         |    1      |     7
ans_tr2_f04      |  AN            |    2      |     4
ans_tr3_f06      |  AN            |    3      |     6
brp_tr1_f09      |  BP            |    1      |     9
brp_tr2_f12      |  BP            |    2      |    12
cn_tr2_f17       |  CN            |    2      |    17
cn_tr1_f15       |  CN            |    1      |    15

**Calculate total basla area for plots**
Total Basal area in meteres squared

**Calculate biomass in grams**

Use allometric equations from Alexander et al (2012) for trees and Berner et al 2015 for shrubs. 
Equation: M=ax^b	[x=BD or DBH]

Genus    |	latin	  | BD 	      |   a	  |    b
---------|----------|-----------|-------|---------
Alder    |	alnus   |	0.18-9.52 |	23.7  |	 2.68
birch    |	betual  |	0.09-2.53	| 28.1	|  2.97
willow   |	salix	  | 0.10-6.30	| 23.53 |	 2.83
larch	   |	        | 0.7-39.2	| 39.46	|  2.26


Genus    |	latin	  |  	 DBH    |   a	  |    b
---------|----------|-----------|-------|---------			
Larch	   |          |	0.08-29.3	| 179.2	|  2.01

# Pakcages
```{r}
library(tidyr)
library(dplyr)
library(stringr)
```

# Functions

```{r}
numextract <- function(string){ 
  str_extract(string, "\\-*\\d+\\.*\\d*")
} 


```

# Site info

```{r}
site_info = read.csv("../data/field/site_info_2018-2019.csv")
```


```{r}
head(site_info)
```

# Subset field data by transect

## Create a unique ID for each point
```{r}
site_info2 = site_info %>% mutate(ID =  paste(site, burn_year, transect, plot, sep = '_')) %>%
  dplyr::select(ID, longitude, latitude, elevation, month:photopoint_files)
site_info2
```


Filter out required field transects that coorespond to UAV data
```{r}
fd_alnus_tr1 = site_info2 %>% group_by(site) %>% filter(site=='Alnus') %>% group_by(transect) %>% filter(transect == 1)
fd_ans_tr2 = site_info2 %>% group_by(site) %>% filter(site=='ANS') %>% group_by(transect) %>% filter(transect == 2)
fd_ans_tr3 = site_info2 %>% group_by(site) %>% filter(site=='ANS') %>% group_by(transect) %>% filter(transect == 3)
fd_bp_tr1 = site_info2 %>% group_by(site) %>% filter(site=='BP') %>% group_by(transect) %>% filter(transect == 1)
fd_bp_tr2 = site_info2 %>% group_by(site) %>% filter(site=='BP') %>% group_by(transect) %>% filter(transect == 2)
fd_cn_tr1 = site_info2 %>% group_by(region) %>% filter(site=='CN') %>% group_by(transect) %>% filter(transect == 1)
fd_cn_tr2 = site_info2 %>% group_by(region) %>% filter(site=='CN') %>% group_by(transect) %>% filter(transect == 2)
```

Recombind filtered transects from above.
```{r}
ans_12 = rbind(fd_ans_tr2 , fd_ans_tr3)
brp_12 = rbind( fd_bp_tr1, fd_bp_tr2)
cn_12 = rbind(fd_cn_tr1, fd_cn_tr2)
alnus_ans = rbind(fd_alnus_tr1, ans_12)
brp_cn = rbind(brp_12, cn_12)
site_info_drones = rbind(alnus_ans, brp_cn)
site_info_drones
```

```{r}
length(unique(site_info_drones$ID))
```

```{r}
head(site_info_drones)
```

 Create a plot_dis column to pull out numeric values from the plot column. There  are some sites that  had two  plots  at  -25 labeled  as -25a and -25b, which is  not useful for analysis.  Unique ID and plot  include the a or b designation
```{r}
sites_subset = site_info_drones %>% mutate(plot_dis = numextract(plot)) %>% mutate(burn_unburn = ifelse(plot_dis>=0, "burned", "unburned")) %>% dplyr::select(ID, region:plot, plot_dis, burn_unburn, slope:aspect) %>% mutate(burn_year = as.factor(burn_year))
sites_subset
```

```{r eval=FALSE, include=FALSE}
write.csv(site_info_drones,"../outputs/field_subset/site_info_uav.csv", row.names = FALSE)
```

# Tree Info

```{r}
tree_info = read.csv("../data/field/tree_survey_2018-2019.csv")
```


```{r}
head(tree_info)
```

# Subset field data by transect

## Create a unique ID for each point
  - and select the columns needed to agregate to plot level
  - create plot_dis column to have distance along transect
  - create unique id for each plot


```{r}
tree_info2 = tree_info %>% mutate(ID =  paste(site, burn_year, transect, plot, sep = '_')) %>% mutate(plot_dis = numextract(plot)) %>%
  dplyr::select(ID, region:plot, plot_dis, diameter_cm:basal_breast, live, radius, area_sampled_m2)
tree_info2
```


```{r}
fd_alnus_tr1 = tree_info2 %>% group_by(site) %>% filter(site=='Alnus') %>% group_by(transect) %>% filter(transect == 1)
fd_ans_tr2 = tree_info2 %>% group_by(site) %>% filter(site=='ANS') %>% group_by(transect) %>% filter(transect == 2)
fd_ans_tr3 = tree_info2 %>% group_by(site) %>% filter(site=='ANS') %>% group_by(transect) %>% filter(transect == 3)
fd_bp_tr1 = tree_info2 %>% group_by(site) %>% filter(site=='BP') %>% group_by(transect) %>% filter(transect == 1)
fd_bp_tr2 = tree_info2 %>% group_by(site) %>% filter(site=='BP') %>% group_by(transect) %>% filter(transect == 2)
fd_cn_tr1 = tree_info2 %>% group_by(region) %>% filter(site=='CN') %>% group_by(transect) %>% filter(transect == 1)
fd_cn_tr2 = tree_info2 %>% group_by(region) %>% filter(site=='CN') %>% group_by(transect) %>% filter(transect == 2)
```

```{r}
ans_12 = rbind(fd_ans_tr2 , fd_ans_tr3)
brp_12 = rbind( fd_bp_tr1, fd_bp_tr2)
cn_12 = rbind(fd_cn_tr1, fd_cn_tr2)
alnus_ans = rbind(fd_alnus_tr1, ans_12)
brp_cn = rbind(brp_12, cn_12)
tree_info_drones = rbind(alnus_ans, brp_cn)
tree_info_drones
```


## Aggregate tree info to plot info
Summarise tree info to plot
  - Convert diamter to basal area 
  - Total basal area
  - count of live stems
  - count of dead stems


### Biomass

Biomass in  grams ber squared meter
Equation: M=ax^b	[x=BD or DBH]

Genus    |	latin	  | BD 	      |   a	  |    b
---------|----------|-----------|-------|---------
larch	   |	        | 0.7-39.2	| 39.46	|  2.26


Genus    |	latin	  |  	 DBH    |   a	  |    b
---------|----------|-----------|-------|---------			
Larch	   |          |	0.08-29.3	| 179.2	|  2.01

For biomass in grams of carbon follow
"Tree biomass was converted to C mass using a C concentration of 46 % for foliage (live trees only), 47 % for stemwood/bark and snag, and 48 % for branches (Alexander et al., 2012)."

Add coefficients for biomass calculation
```{r}
tree_info_drones = tree_info_drones %>% dplyr::mutate(a = ifelse(basal_breast == "dbh", 179.2,
                                     ifelse(basal_breast == "bd", 39.46, 0))) %>% 
  dplyr::mutate(b = ifelse(basal_breast == "dbh", 2.01,
                                     ifelse(basal_breast == "bd", 2.26, 0)))

tree_info_drones
```

Filter for live trees only

```{r}
tree_info_drones_live = tree_info_drones %>% filter(live=="live")
tree_info_drones_live
```


Calculate all tree biomass
```{r}
tree_biomass = tree_info_drones %>% dplyr::mutate(tree_biomass = a*diameter_cm^b)
tree_biomass
```

Biomass by plot

```{r}
tree_total_biomass = tree_biomass %>%
    group_by(ID, site, transect, plot, live) %>%
    summarise(tree_total_biomass = sum(tree_biomass)) %>% 
    spread(live, tree_total_biomass, fill = 0) %>%
    rename(live_tree_biomass = live, dead_tree_biomass = dead) %>%
    mutate(tree_total_biomass = live_tree_biomass + dead_tree_biomass) %>%
    mutate(tree_total_biomass_C = tree_total_biomass*.46) %>%
    mutate(live_tree_biomass_c = live_tree_biomass*.46) %>%
    mutate(dead_tree_biomass_c = dead_tree_biomass*.46)
    
tree_total_biomass  
```


### Basal Area

basal area for each tree
```{r}
tree_ba = dplyr::mutate(tree_info_drones, tree_ba = ((diameter_cm^2)*0.00007854))
```



Basal area per plot

```{r}
tree_total_ba = tree_ba %>%
    group_by(ID, site, transect, plot, live) %>%
    summarise(tree_total_ba = sum(tree_ba)) %>%
    spread(live, tree_total_ba, fill = 0) %>%
    mutate(tree_total_ba = live + dead) %>%
    rename(live_tree_ba = live, dead_tree_ba = dead)
tree_total_ba  
```


### Tree plot area
```{r}
tree_plot = tree_info_drones %>%
    group_by(ID, site, transect, plot) %>%
    summarise(tree_plot_area = median(area_sampled_m2)) 

tree_plot  
```
### tree radius
```{r}
tree_radius = tree_info_drones %>%
    group_by(ID, site, transect, plot) %>%
    summarise(tree_radius = median(radius))

tree_radius  
```


```{r}
stems = tree_info_drones %>%
    group_by(ID, site, transect, plot, live) %>%
    summarise(n = n()) %>%
    mutate(stems = sum(n)) %>%
    spread(live, n, fill = 0)
stems
```
### Merge tree data
```{r}
plot_radius = merge(tree_plot, tree_radius,  by=c("ID","site", "transect", "plot"))
plot_radius
```

```{r}
stems_ba = merge(stems,  tree_total_ba,  by=c("ID","site", "transect", "plot"))
stems_ba
```


```{r}
tree_info_uav = merge(plot_radius,  stems_ba,  by=c("ID","site", "transect", "plot"))
tree_info_uav
```

```{r}
tree_subset = merge(tree_info_uav,  tree_total_biomass,  by=c("ID","site", "transect", "plot"))
tree_subset
```


```{r eval=FALSE, include=FALSE}
write.csv(tree_subset,"../outputs/field_subset/tree_plot.csv", row.names = FALSE)
```


### Create a file for plot radius for tree data

```{r}
tree_radius = tree_subset %>% dplyr::select(ID, tree_plot_area:tree_radius)
```

```{r eval=FALSE, include=FALSE}
write.csv(tree_radius,"../outputs/field_subset/tree_radius.csv", row.names = FALSE)
```

# Shrub Info

```{r}
shrub_info = read.csv("../data/field/shrubs_2018-2019.csv")
```

```{r}
summary(shrub_info)
```

```{r}
head(shrub_info)
```

# Subset field data by transect
```{r}
shrub_info %>% mutate(genus = word(species, 1))
```

## Create a unique ID for each point
  - create plot_dis column to have distance along transect
  - create unique id for each plot
  - create column for genus for biomass equation
  - and select the columns needed to agregate to plot level
```{r}
shrub_info2 = shrub_info %>% mutate(ID =  paste(site, burn_year, transect, plot, sep = '_')) %>% mutate(plot_dis = numextract(plot)) %>% mutate(genus = word(species, 1)) %>%
 dplyr:: select(ID, region:plot, plot_dis, plot_width:species, genus, basal_diameter_cm)
shrub_info2
```


```{r}
fd_alnus_tr1 = shrub_info2 %>% group_by(site) %>% filter(site=='Alnus') %>% group_by(transect) %>% filter(transect == 1)
fd_ans_tr2 = shrub_info2 %>% group_by(site) %>% filter(site=='ANS') %>% group_by(transect) %>% filter(transect == 2)
fd_ans_tr3 = shrub_info2 %>% group_by(site) %>% filter(site=='ANS') %>% group_by(transect) %>% filter(transect == 3)
fd_bp_tr1 = shrub_info2 %>% group_by(site) %>% filter(site=='BP') %>% group_by(transect) %>% filter(transect == 1)
fd_bp_tr2 = shrub_info2 %>% group_by(site) %>% filter(site=='BP') %>% group_by(transect) %>% filter(transect == 2)
fd_cn_tr1 = shrub_info2 %>% group_by(region) %>% filter(site=='CN') %>% group_by(transect) %>% filter(transect == 1)
fd_cn_tr2 = shrub_info2 %>% group_by(region) %>% filter(site=='CN') %>% group_by(transect) %>% filter(transect == 2)
```

```{r}
ans_12 = rbind(fd_ans_tr2 , fd_ans_tr3)
brp_12 = rbind( fd_bp_tr1, fd_bp_tr2)
cn_12 = rbind(fd_cn_tr1, fd_cn_tr2)
alnus_ans = rbind(fd_alnus_tr1, ans_12)
brp_cn = rbind(brp_12, cn_12)
shrub_info_drones = rbind(alnus_ans, brp_cn)
shrub_info_drones
```

## Aggregate tree info to plot info
Summarise tree info to plot
  - calculate biomass by genus
  - Convert diamter to basal area (BA = 0.005454*DBH^2)
  - Total basal area
  - count of live stems
  - count of dead stems

### Biomass

Biomass as grams per square meter calculated as
Equation: M=ax^b	[x=BD or DBH]

Genus    |	latin	  | BD 	      |   a	  |    b
---------|----------|-----------|-------|---------
Alder    |	alnus   |	0.18-9.52 |	23.7  |	 2.68
birch    |	betual  |	0.09-2.53	| 28.1	|  2.97
willow   |	salix	  | 0.10-6.30	| 23.53 |	 2.83

For biomass as grams of carbon per square meter
"We converted live understory biomass values to C pools by multiplying biomass by 48 % C content  (webb et al. 2017)."

Add coefficients for biomass calculation
```{r}
shrub_info_drones = shrub_info_drones %>% dplyr::mutate(a = ifelse(genus == "Alnus", 23.7,
                                     ifelse(genus == "Betula", 28.1,
                                            ifelse(genus == "Salix", 23.53, 0)))) %>% 
  dplyr::mutate(b = ifelse(genus == "Alnus", 2.68,
                                     ifelse(genus == "Betula", 2.97,
                                            ifelse(genus == "Salix", 2.83, 0))))

shrub_info_drones
```

Calculate biomass per shrub
```{r}
shrub_biomass = shrub_info_drones %>% dplyr::mutate(shrub_biomass = a*basal_diameter_cm^b)
shrub_biomass
```

Calculate total shrub biomass per plot

```{r}
shrub_total_biomass = shrub_biomass %>%
    group_by(ID, site, transect, plot) %>%
    summarise(shrub_total_biomass = sum(shrub_biomass)) %>%
  mutate(shrub_total_biomass_C = shrub_total_biomass*.48)
shrub_total_biomass  
```

###  BA per shrub
```{r}
shrub_ba = dplyr::mutate(shrub_info_drones, shrub_ba = ((basal_diameter_cm^2)*0.00007854))
shrub_ba
```
```{r}
shrub_total_ba = shrub_ba  %>%
    group_by(ID, site, transect, plot) %>%
    summarise(shrub_total_ba = sum(shrub_ba))

shrub_total_ba  
```

```{r}
genus_ba = shrub_ba %>%
    group_by(ID, site, transect, plot, genus) %>%
    summarise(shrub_total_ba = sum(shrub_ba)) %>%
    spread(genus, shrub_total_ba, fill = 0) %>% rename(alnus_ba = Alnus, betual_ba = Betula, salix_ba = Salix)
genus_ba  
```

### Plot size

```{r}
plot_area = dplyr::mutate(shrub_info_drones, plot_area = (plot_width*plot_length)) 

plot_area
```
```{r}
shrub_plot_area = plot_area  %>%
    group_by(ID, site, transect, plot) %>%
    summarise(shrub_plot_area = median(plot_area))

shrub_plot_area  
```


```{r}
shrub_width_plot = shrub_info_drones %>%
    group_by(ID, site, transect, plot) %>%
    summarise(shrub_plot_width = median(plot_width)) 

shrub_width_plot  
```

```{r}
shrub_length_plot = shrub_info_drones %>%
    group_by(ID, site, transect, plot) %>%
    summarise(shrub_plot_length = median(plot_length)) 

shrub_length_plot  
```

### Count by genus
```{r}
shrub_genus_ct = shrub_info_drones %>%
    group_by(ID, site, transect, plot, genus) %>%
    summarise(n = n()) %>%
    mutate(shrub_ct = sum(n)) %>%
    spread(genus, n, fill = 0) %>%
    rename(alnus_ct = Alnus, betual_ct = Betula, salix_ct = Salix)
shrub_genus_ct
```




### combine
```{r}
merge1 = merge(shrub_total_biomass, shrub_total_ba,  by=c("ID","site", "transect", "plot"))
merge1
```

```{r}
merge2 = merge(genus_ba,  shrub_genus_ct,  by=c("ID","site", "transect", "plot"))
merge2
```

```{r}
merge3 = merge(shrub_width_plot, shrub_length_plot, by=c("ID","site", "transect", "plot"))
```

```{r}
merge3_5 = merge(shrub_plot_area, merge3, by=c("ID","site", "transect", "plot"))
```

```{r}
merge4 = merge(merge1,  merge2,  by=c("ID","site", "transect", "plot"))
merge4
```
```{r}
shrub_subset = merge(merge4,  merge3_5,  by=c("ID","site", "transect", "plot"))
shrub_subset
```




```{r eval=FALSE, include=FALSE}
write.csv(shrub_subset,"../outputs/field_subset/shrub_plot.csv", row.names = FALSE)
```

```{r}
shrub_plot_size = shrub_subset %>% dplyr::select(ID, shrub_plot_area, shrub_plot_width, shrub_plot_length)
```


```{r eval=FALSE, include=FALSE}
write.csv(shrub_plot_size,"../outputs/field_subset/shrub_plot_size.csv", row.names = FALSE)
```


# Combine site, tree, and shrub data

```{r}
com_merge1 = merge(tree_subset, shrub_subset,  by=c("ID","site", "transect", "plot"))
field_data = merge(sites_subset, com_merge1,  by=c("ID","site", "transect", "plot"))
field_data
```
```{r}
field_data = field_data %>% 
  mutate(transect_ID =  paste(site, burn_year, transect, sep = '_'))  %>%
  mutate(prop_live = (live/stems*100)) %>%
  mutate(woody_ba = (tree_total_ba + shrub_total_ba)) %>% 
  
  mutate(shrub_ba_m2 =  shrub_total_ba/shrub_plot_area) %>%
  mutate(shrub_ba_314m2 = shrub_ba_m2*314 ) %>%
  mutate(shrub_ba_ha = shrub_ba_m2*10000) %>%
  mutate(shrub_ba_100m2 = shrub_ba_m2*100) %>%
  
  mutate(live_tree_ba_m2 =  live_tree_ba/tree_plot_area) %>%
  mutate(live_tree_ba_314m2 = live_tree_ba_m2*314 ) %>%
  mutate(live_tree_ba_ha = live_tree_ba_m2*10000) %>%
  mutate(live_tree_ba_100m2 = live_tree_ba_m2*100) %>%
  
  mutate(dead_tree_ba_m2 =  dead_tree_ba/tree_plot_area) %>%
  mutate(dead_tree_ba_314m2 = dead_tree_ba_m2*314 ) %>%
  mutate(dead_tree_ba_ha = dead_tree_ba_m2*10000) %>%
  mutate(dead_tree_ba_100m2 = dead_tree_ba_m2*100) %>%
  
  mutate(tree_biomass_g_m = tree_total_biomass/tree_plot_area) %>% 
  mutate(tree_biomass_C_g_m = tree_total_biomass_C/tree_plot_area) %>%
  
  mutate(live_tree_biomass_g_m = live_tree_biomass/tree_plot_area) %>% 
  mutate(live_tree_biomass_C_g_m = live_tree_biomass_c/tree_plot_area) %>%
  
  mutate(shrub_biomass_g_m = shrub_total_biomass/shrub_plot_area) %>% 
  mutate(shrub_biomass_C_g_m = shrub_total_biomass_C/shrub_plot_area) %>% 
  
  mutate(woody_biomass = (tree_total_biomass + shrub_total_biomass)) %>% 
  mutate(woody_biomass_g_m = tree_biomass_g_m + shrub_biomass_g_m) %>%
  mutate(woody_total_biomass_C = (tree_total_biomass_C + shrub_total_biomass_C)) %>%
  mutate(woody_biomass_C_g_m = tree_biomass_C_g_m + shrub_biomass_C_g_m) %>%
 
  
  
  mutate(live_woody_biomass = (live_tree_biomass + shrub_total_biomass)) %>% 
  mutate(live_woody_biomass_g_m = live_tree_biomass_g_m + shrub_biomass_g_m) %>%
  mutate(live_woodytotal_biomass_C = (live_tree_biomass_c + shrub_total_biomass_C)) %>%
  mutate(live_woody_biomass_C_g_m = live_tree_biomass_C_g_m + shrub_biomass_C_g_m) 
  
  

field_data

```

```{r eval=FALSE, include=FALSE}
write.csv(field_data,"../outputs/field_subset/field_data_all.csv", row.names = FALSE)
```

# Field data summary ranges

Prop Live
```{r}
min(field_data$prop_live)
max(field_data$prop_live)
```

Tree Basal Area
```{r}
min(field_data$tree_total_ba)
max(field_data$tree_total_ba)
```

```{r}
min(field_data$live_tree_ba)
max(field_data$live_tree_ba)
```

```{r}
min(field_data$live_tree_ba_ha)
max(field_data$live_tree_ba_ha)
```

shrub Basal Area
```{r}
min(field_data$shrub_total_ba)
max(field_data$shrub_total_ba)
```

shrub Basal Area
```{r}
min(field_data$shrub_ba_ha)
max(field_data$shrub_ba_ha)
```
