==> scripts/1.2_Field-data-spatial.Rmd <==
---
title: "UntitledField data sort and buffer"
author: "Anna Talucci"
date: "2/24/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```



```{r, include=FALSE}
library(sp) # spatial data
library(rgdal) # geospatial data
library(raster) # deals with rasters
library(ggplot2)
library(rgeos) # GIS Geometry
library(foreign) # deals with dbf files
library(tidyr)
library(dplyr)
library(latticeExtra)
library(sf)
```

# Projections

WGS 84 need for gee
```{r}
wgs_proj =  "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs "

```

Equal Area projection for russia need for buffering hotspot data
https://gis.stackexchange.com/questions/264093/projection-for-russia-map

+proj=aea +lat_1=50 +lat_2=70 +lat_0=56 +lon_0=100 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs 
+proj=aea +lat_1=50 +lat_2=70 +lat_0=56 +lon_0=100 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0 

```{r}
ee_russia = "+proj=aea +lat_1=50 +lat_2=70 +lat_0=56 +lon_0=100 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0 "
```


# Field data

```{r}
field_data = read.csv("../data/field/site_info_2018-2019.csv")
tree_data = read.csv("../data/field_subset/tree_radius.csv")
shrub_data = read.csv("../data/field_subset/shrub_plot_size.csv")
```


```{r}
head(field_data)
```

```{r}
head(tree_data)
```

```{r}
head(shrub_data)
```
# Subset field data by transect

## Create a unique ID for each point
```{r}
field_data3 = field_data %>% mutate(ID =  paste(site, burn_year, transect, plot, sep = '_')) %>%
  dplyr::select(ID, longitude:elevation, region:plot, slope:aspect)
field_data3
```

```{r eval=FALSE, include=FALSE}
merge1 = merge(field_data2, shrub_data, by="ID")
merge2 = merge(merge1, tree_data, by="ID")

field_data3 = merge2 %>% rename(shrub_area = area_sampled_m2.x, area_tree = area_sampled_m2.y)
field_data3
```



## ANS_tr2

Sort and filter field data transects
```{r}
fd_ans_tr2 = field_data3 %>% group_by(site) %>% filter(site=='ANS') %>% group_by(transect) %>% filter(transect == 2)
fd_ans_tr2
```


Create spatial points data frame with attributes
Define Lat/Long
```{r}
coords_ans_tr2 = cbind(fd_ans_tr2$longitude, fd_ans_tr2$latitude)

fd_ans_tr2_pt <- SpatialPointsDataFrame(coords_ans_tr2, proj4string = CRS(wgs_proj), data = fd_ans_tr2)
```

Transform to Equal area projection
```{r}
fd_ans_tr2_ee <- spTransform(fd_ans_tr2_pt, CRS(ee_russia))
```

Generate a spatial buffer
```{r}
fd_ans_tr2_ee_10m = gBuffer(fd_ans_tr2_ee, width = 10, capStyle="ROUND", joinStyle="ROUND")
```

Transform from sp to sf
```{r}
fd_ans_tr2_ee_sf = st_as_sf(fd_ans_tr2_ee)
fd_ans_tr2_ee_10m_sf = st_as_sf(fd_ans_tr2_ee_10m)
```



Plot with ggplot and labels
```{r eval=FALSE, include=FALSE}
ggplot() +
  geom_sf(data=fd_ans_tr2_pt_sf) +
  geom_text(data = fd_ans_tr2_pt_sf, aes(x = longitude, y=latitude, label = ID), size = 1)
```

## ANS_tr3
Sort and filter field data transects
```{r}
fd_ans_tr3 = field_data3 %>% group_by(site) %>% filter(site=='ANS') %>% group_by(transect) %>% filter(transect == 3)
fd_ans_tr3
```


Create spatial points data frame with attributes
```{r}
coords_ans_tr3 = cbind(fd_ans_tr3$longitude, fd_ans_tr3$latitude)
fd_ans_tr3_pt <- SpatialPointsDataFrame(coords_ans_tr3, proj4string = CRS(wgs_proj), data = fd_ans_tr3)
```

Transform to Equal area projection
```{r}
fd_ans_tr3_ee <- spTransform(fd_ans_tr3_pt, CRS(ee_russia))
```

Generate a spatial buffer
```{r}
fd_ans_tr3_ee_10m = gBuffer(fd_ans_tr3_ee, width = 10, capStyle="ROUND", joinStyle="ROUND")
```

Transform from sp to sf
```{r}
fd_ans_tr3_ee_sf = st_as_sf(fd_ans_tr3_ee )
fd_ans_tr3_ee_10m_sf = st_as_sf(fd_ans_tr3_ee_10m)
```





## alnus_tr1

Sort and filter field data transects
```{r}
fd_alnus_tr1 = field_data3 %>% group_by(site) %>% filter(site=='Alnus') %>% group_by(transect) %>% filter(transect == 1)
fd_alnus_tr1
```

Sort and filter field data transects
```{r}
coords_alnus_tr1 = cbind(fd_alnus_tr1$longitude, fd_alnus_tr1$latitude)
fd_alnus_tr1_pt <- SpatialPointsDataFrame(coords_alnus_tr1, proj4string = CRS(wgs_proj), data = fd_alnus_tr1)
```

Transform to Equal area projection
```{r}
fd_alnus_tr1_ee <- spTransform(fd_alnus_tr1_pt, CRS(ee_russia))
```

Generate a spatial buffer
```{r}
fd_alnus_tr1_ee_10m = gBuffer(fd_alnus_tr1_ee, width = 10, capStyle="ROUND", joinStyle="ROUND")
```

Transform from sp to sf
```{r}
fd_alnus_tr1_ee_sf = st_as_sf(fd_alnus_tr1_ee )
fd_alnus_tr1_ee_10m_sf = st_as_sf(fd_alnus_tr1_ee_10m)
```



## BP_tr1

Sort and filter field data transects
```{r}
fd_bp_tr1 = field_data3 %>% group_by(site) %>% filter(site=='BP') %>% group_by(transect) %>% filter(transect == 1)
fd_bp_tr1
```

Define coordinates and generate spatial point data frame with attributes
```{r}
coords_bp_tr1 = cbind(fd_bp_tr1$longitude, fd_bp_tr1$latitude)
fd_bp_tr1_pt <- SpatialPointsDataFrame(coords_bp_tr1, proj4string = CRS(wgs_proj), data = fd_bp_tr1)
```

Transform to Equal area projection
```{r}
fd_bp_tr1_ee <- spTransform(fd_bp_tr1_pt, CRS(ee_russia))
```

Generate a spatial buffer
```{r}
fd_bp_tr1_ee_10m = gBuffer(fd_bp_tr1_ee, width = 10, capStyle="ROUND", joinStyle="ROUND")
```

Transform from sp to sf
```{r}
fd_bp_tr1_ee_sf = st_as_sf(fd_bp_tr1_ee)
fd_bp_tr1_ee_10m_sf = st_as_sf(fd_bp_tr1_ee_10m)
```





## BP_tr2

Sort and filter field data transects
```{r}
fd_bp_tr2 = field_data3 %>% group_by(site) %>% filter(site=='BP') %>% group_by(transect) %>% filter(transect == 2)
fd_bp_tr2
```

Define coordinates and generate spatial point data frame with attributes
```{r}
coords_bp_tr2 = cbind(fd_bp_tr2$longitude, fd_bp_tr2$latitude)
fd_bp_tr2_pt <- SpatialPointsDataFrame(coords_bp_tr2 , proj4string = CRS(wgs_proj), data = fd_bp_tr2)
```

Transform to Equal area projection
```{r}
fd_bp_tr2_ee <- spTransform(fd_bp_tr2_pt, CRS(ee_russia))
```

Generate a spatial buffer
```{r}
fd_bp_tr2_ee_10m = gBuffer(fd_bp_tr2_ee, width = 10, capStyle="ROUND", joinStyle="ROUND")
```

Transform from sp to sf
```{r}
fd_bp_tr2_ee_sf = st_as_sf(fd_bp_tr2_ee)
fd_bp_tr2_ee_10m_sf = st_as_sf(fd_bp_tr2_ee_10m)
```



## CN_tr1
Sort and filter field data transects
```{r}
fd_cn_tr1 = field_data3 %>% group_by(region) %>% filter(region == "Cherskiy") %>% group_by(site) %>% filter(site=='CN') %>% group_by(transect) %>% filter(transect == 1)
fd_cn_tr1
```


Define coordinates and generate spatial point data frame with attributes
```{r}
coord_cn_tr1 = cbind(fd_cn_tr1$longitude, fd_cn_tr1$latitude)
fd_cn_tr1_pt <- SpatialPointsDataFrame(coord_cn_tr1, proj4string = CRS(wgs_proj), data = fd_cn_tr1)
```

Transform to Equal area projection
```{r}
fd_cn_tr1_ee <- spTransform(fd_cn_tr1_pt, CRS(ee_russia))
```

Generate a spatial buffer
```{r}
fd_cn_tr1_ee_10m = gBuffer(fd_cn_tr1_ee, width = 10, capStyle="ROUND", joinStyle="ROUND")
```

Transform from sp to sf
```{r}
fd_cn_tr1_ee_sf = st_as_sf(fd_cn_tr1_ee)
fd_cn_tr1_ee_10m_sf = st_as_sf(fd_cn_tr1_ee_10m)
```




## CN_tr2
Sort and filter field data transects
```{r}
fd_cn_tr2 = field_data3 %>% group_by(region) %>% filter(region == "Cherskiy") %>% group_by(site) %>% filter(site=='CN') %>% group_by(transect) %>% filter(transect == 2)
fd_cn_tr2
```

Define coordinates and generate spatial point data frame with attributes
```{r}
coords_cn_tr2 = cbind(fd_cn_tr2$longitude, fd_cn_tr2$latitude)
fd_cn_tr2_pt <- SpatialPointsDataFrame(coords_cn_tr2, proj4string = CRS(wgs_proj), data = fd_cn_tr2)
```

Transform to Equal area projection
```{r}
fd_cn_tr2_ee <- spTransform(fd_cn_tr2_pt, CRS(ee_russia))
```

Generate a spatial buffer
```{r}
fd_cn_tr2_ee_10m = gBuffer(fd_cn_tr2_ee, width = 10, capStyle="ROUND", joinStyle="ROUND")
```

Transform from sp to sf
```{r}
fd_cn_tr2_ee_sf = st_as_sf(fd_cn_tr2_ee)
fd_cn_tr2_ee_10m_sf = st_as_sf(fd_cn_tr2_ee_10m)
```




## Write to shape file

Points
```{r eval=FALSE, include=FALSE}
st_write(fd_ans_tr2_ee_sf, "../outputs/field-data/point/ans_tr2_ee_pt.shp", driver="ESRI Shapefile") 
st_write(fd_ans_tr3_ee_sf, "../outputs/field-data/point/ans_tr3_ee_pt.shp", driver="ESRI Shapefile") 
st_write(fd_alnus_tr1_ee_sf, "../outputs/field-data/point/alnus_tr1_ee_pt.shp", driver="ESRI Shapefile")
st_write(fd_bp_tr1_ee_sf, "../outputs/field-data/point/bp_tr1_ee_pt.shp", driver="ESRI Shapefile") 
st_write(fd_bp_tr2_ee_sf, "../outputs/field-data/point/bp_tr2_ee_pt.shp", driver="ESRI Shapefile") 
st_write(fd_cn_tr1_ee_sf, "../outputs/field-data/point/cn_tr1_ee_pt.shp", driver="ESRI Shapefile") 
st_write(fd_cn_tr2_ee_sf, "../outputs/field-data/point/cn_tr2_ee_pt.shp", driver="ESRI Shapefile") 

```

Buffers
```{r eval=FALSE, include=FALSE}
st_write(fd_ans_tr2_ee_10m_sf, "../outputs/field-data/buffer/ans_tr2_ee_10m.shp", driver="ESRI Shapefile") 
st_write(fd_ans_tr3_ee_10m_sf, "../outputs/field-data/buffer/ans_tr3_ee_10m.shp", driver="ESRI Shapefile") 
st_write(fd_alnus_tr1_ee_10m_sf, "../outputs/field-data/buffer/alnus_tr1_ee_10m.shp", driver="ESRI Shapefile") 
st_write(fd_bp_tr1_ee_10m_sf, "../outputs/field-data/buffer/bp_tr1_ee_10m.shp", driver="ESRI Shapefile") 
st_write(fd_bp_tr2_ee_10m_sf, "../outputs/field-data/buffer/bp_tr2_ee_10m.shp", driver="ESRI Shapefile") 
st_write(fd_cn_tr1_ee_10m_sf, "../outputs/field-data/buffer/cn_tr1_ee_10m.shp", driver="ESRI Shapefile") 
st_write(fd_cn_tr2_ee_10m_sf , "../outputs/field-data/buffer/cn_tr2_ee_10m.shp", driver="ESRI Shapefile") 
 
```

# Create  shape file  with  only the middle plot for mapping figure
```{r}
ans = rbind(fd_ans_tr2, fd_ans_tr3)
ans_alnus = rbind(ans, fd_alnus_tr1)
bp = rbind(fd_bp_tr1, fd_bp_tr2)
cn = rbind(fd_cn_tr1, fd_cn_tr2)
ans_alnus_bp  = rbind(ans_alnus, bp)
ans_alnus_bp_cn = rbind(ans_alnus_bp, cn)
ans_alnus_bp_cn
```

```{r}
plot100 = ans_alnus_bp_cn %>% filter(plot== 100) %>% dplyr::select(ID:latitude, region:plot)
```

Define coordinates and generate spatial point data frame with attributes
```{r}
coords_plot100 = cbind(plot100$longitude, plot100$latitude)
plot100_pt <- SpatialPointsDataFrame(coords_plot100, proj4string = CRS(wgs_proj), data = plot100)
```

Transform from sp to sf
```{r}
plot100_pt_sf = st_as_sf(plot100_pt)
```

```{r}
st_write(plot100_pt_sf, "../outputs/field-data/point/plot100_pt_sf.shp", driver="ESRI Shapefile") 
```

## Manual select coords for each transect for map

```{r}
ans_alnus_bp_cn
```

```{r}
ANS_2003_2_100 = ans_alnus_bp_cn %>% filter(ID== "ANS_2003_2_100") 
ANS_2003_3_100 = ans_alnus_bp_cn %>% filter(ID== "ANS_2003_3_100") 
Alnus_1984_1_100 = ans_alnus_bp_cn %>% filter(ID== "Alnus_1984_1_100") 
BP_1983_1_25 = ans_alnus_bp_cn %>% filter(ID== "BP_1983_1_25") 
BP_1983_2_200 = ans_alnus_bp_cn %>% filter(ID== "BP_1983_2_200") 
CN_2001_1_25 = ans_alnus_bp_cn %>% filter(ID== "CN_2001_1_25") 
CN_2001_2_200 = ans_alnus_bp_cn %>% filter(ID== "CN_2001_2_200") 
```

```{r}
ans = rbind(ANS_2003_2_100, ANS_2003_3_100)
ans_al = rbind(ans, Alnus_1984_1_100)
bp = rbind(BP_1983_1_25, BP_1983_2_200)
cn = rbind(CN_2001_1_25, CN_2001_2_200)
ans_al_bp = rbind(ans_al, bp)
ans_al_bp_cn = rbind(ans_al_bp, cn)
```


```{r}
plot_spot = ans_al_bp_cn %>% dplyr::select(ID:latitude, region:plot)
```

Define coordinates and generate spatial point data frame with attributes
```{r}
coords_plot_spot = cbind(plot_spot$longitude, plot_spot$latitude)
plot_spot_pt <- SpatialPointsDataFrame(coords_plot_spot, proj4string = CRS(wgs_proj), data = plot_spot)
```

Transform from sp to sf
```{r}
plot_spot_pt_sf = st_as_sf(plot_spot_pt)
```

```{r}
st_write(plot_spot_pt_sf, "../outputs/field-data/point/plot_transect_map_pt.shp", driver="ESRI Shapefile") 
```


Define coordinates and generate spatial point data frame with attributes
```{r}
coords_plot_all = cbind(ans_al_bp_cn$longitude, ans_al_bp_cn$latitude)
plot_all_pt <- SpatialPointsDataFrame(coords_plot_all, proj4string = CRS(wgs_proj), data = ans_al_bp_cn)
```

Transform from sp to sf
```{r}
plot_all_pt_sf = st_as_sf(plot_all_pt)
```

```{r}
st_write(plot_all_pt_sf, "../outputs/field-data/point/plot_all_transect_map_pt.shp", driver="ESRI Shapefile") 
```
