==> scripts/2.1_Calculate-GCC-NDVI.Rmd <==
---
title: "Calculate GCC and NDVI"
author: "Anna Talucci"
date: "February 7, 2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r, include=FALSE}
library(sp) # spatial data
library(rgdal) # geospatial data
library(raster) # deals with rasters
library(rgeos) # GIS Geometry
library(foreign) # deals with dbf files
library(dplyr)
library(latticeExtra)
```


# RGB rasters

Stack rasters for gcc calculation

```{r}
alnus_tr1_f07 = stack("../data/rgb/ALN_TR1_FL007.tif")
ans_tr2_f04 = stack("../data/rgb/ALN_TR2_FL004.tif")
ans_tr3_f06 = stack("../data/rgb/ALN_TR3_FL006.tif")
brp_tr1_f09 = stack("../data/rgb/BRP_TR1_FL009.tif")
brp_tr2_f12 = stack("../data/rgb/BRP_TR2_FL012.tif")
cn_tr2_f17 = stack("../data/rgb/CYN_TR2_FL017.tif")
cn_tr1_f15 = stack("../data/rgb/CYN_TR1_FL015.tif")
```

View raster meta data
```{r}
alnus_tr1_f07
ans_tr2_f04
ans_tr3_f06
brp_tr1_f09
brp_tr2_f12
cn_tr2_f17
cn_tr1_f15 
```


# Elena's code
masking to elimate unwanted values 
f1b <- mask(f1[[1:3]],f1[[4]],maskvalue=0)
plot(f1b)

calculating gcc
gcc <- f1b[[2]]/(f1b[[1]]+f1b[[2]]+f1b[[3]])
plot(gcc)

# Look at band details
```{r}
alnus_tr1_f07@layers
```
# Calculate GCC
gcc = G /(R+G+B)
Band 1 = Red
Band 2 = Green
Band 3 = Blue
```{r}
f_gcc <- function(x) {
  (x[,2])/(x[,1]+x[,2]+x[,3])
}
```

alnus_tr1_f07
ans_tr2_f04
ans_tr3_f06
brp_tr1_f09
brp_tr2_f12
cn_tr2_f17
cn_tr1_f15 

cn_f20_ndvi = (cn_f20_4band[[2]] - cn_f20_4band[[3]]) / (cn_f20_4band[[2]] + cn_f20_4band[[3]])

```{r}
alnus_tr1_f07_gcc <- calc(alnus_tr1_f07, f_gcc)
ans_tr2_f04_gcc <- calc(ans_tr2_f04, f_gcc)
ans_tr3_f06_gcc <- calc(ans_tr3_f06, f_gcc)
brp_tr1_f09_gcc <- calc(brp_tr1_f09, f_gcc)
brp_tr2_f12_gcc <- calc(brp_tr2_f12, f_gcc)
cn_tr2_f17_gcc <- calc(cn_tr2_f17, f_gcc)
cn_tr1_f15_gcc <- calc(cn_tr1_f15, f_gcc)
```


```{r}
alnus_tr1_f07_gcc1 <-  (alnus_tr1_f07[[2]] / (alnus_tr1_f07[[1]] + alnus_tr1_f07[[2]] + alnus_tr1_f07[[3]]))

```

```{r}
plot(alnus_tr1_f07_gcc)
```

```{r}
plot(alnus_tr1_f07_gcc1)
```


# Write GCC to file
```{r eval=FALSE, include=FALSE}
writeRaster(alnus_tr1_f07_gcc, filename="../outputs/gcc/alnus_tr1_f07_gcc", format="GTiff", overwrite=TRUE)
writeRaster(ans_tr2_f04_gcc, filename="../outputs/gcc/ans_tr2_f04_gcc", format="GTiff", overwrite=TRUE)
writeRaster(ans_tr3_f06_gcc, filename="../outputs/gcc/ans_tr3_f06_gcc", format="GTiff", overwrite=TRUE)
writeRaster(brp_tr1_f09_gcc, filename="../outputs/gcc/brp_tr1_f09_gcc", format="GTiff", overwrite=TRUE)
writeRaster(brp_tr2_f12_gcc, filename="../outputs/gcc/brp_tr2_f12_gcc", format="GTiff", overwrite=TRUE)
writeRaster(cn_tr2_f17_gcc, filename="../outputs/gcc/cn_tr2_f17_gcc", format="GTiff", overwrite=TRUE)
writeRaster(cn_tr1_f15_gcc, filename="../outputs/gcc/cn_tr1_f15_gcc", format="GTiff", overwrite=TRUE)
```


# Plot GCC images
```{r}
plot(alnus_tr1_f07_gcc)
```

```{r}
plot(ans_tr2_f04_gcc)
```

```{r}
plot(ans_tr3_f06_gcc)
```

```{r}
plot(brp_tr1_f09_gcc)
```

```{r}
plot(brp_tr2_f12_gcc)
```

```{r}
plot(cn_tr2_f17_gcc)
```

# Multispectral Data
```{r}
alnus_f8 <- list.files(path="../data/multispectral/RU_ALN_TR1_FL008M", pattern =".tif$", full.names=TRUE, recursive=TRUE)
cn_f17 <- list.files(path="../data/multispectral/RU_CYN_TR2_FL017B", pattern =".tif$", full.names=TRUE, recursive=TRUE)
cn_f20 <- list.files(path="../data/multispectral/RU_CYN_TR1_FL020M", pattern =".tif$", full.names=TRUE, recursive=TRUE)
```


```{r}
alnus_f8
```


```{r}
alnus_f08_multi = stack(alnus_f8)
cn_f17_multi = stack(cn_f17)
cn_f20_multi = stack(cn_f20)
```


```{r}
alnus_f08_multi
cn_f17_multi 
cn_f20_multi 
```


```{r}
alnus_f08_multi@layers
```
```{r}
plot(alnus_f08_multi[[7]])
```
```{r}
plot(alnus_f08_multi[[8]])
```

## Pull out the 4 bands with data
```{r}
alnus_f08_4band <- alnus_f08_multi[[c(1, 3, 5, 7)]]
cn_f17_4band = cn_f17_multi[[c(1, 3, 5, 7)]] 
cn_f20_4band = cn_f20_multi[[c(1, 3, 5, 7)]] 

```

```{r}
alnus_f08_4band@layers
cn_f17_4band@layers
cn_f17_4band@layers
```


## Select bands to calculated NDVI 
 (nir-red) / (nir+red)
 Band 1 - Green
 Band 2 = NIR
 Band 3 = Red 
 Band 4 = Red Edge

```{r}
alnus_f08_ndvi <- (alnus_f08_4band[[2]] - alnus_f08_4band[[3]]) / (alnus_f08_4band[[2]] + alnus_f08_4band[[3]])
cn_f17_ndvi = (cn_f17_4band[[2]] - cn_f17_4band[[3]]) / (cn_f17_4band[[2]] + cn_f17_4band[[3]])
cn_f20_ndvi = (cn_f20_4band[[2]] - cn_f20_4band[[3]]) / (cn_f20_4band[[2]] + cn_f20_4band[[3]])
```



```{r}
plot(alnus_f08_ndvi)
```

```{r}
plot(cn_f17_ndvi)
```

```{r}
plot(cn_f20_ndvi)
```

### Write to file
```{r eval=FALSE, include=FALSE}
writeRaster(alnus_f08_ndvi, filename="../outputs/ndvi/alnus_f08_NDVI", format="GTiff", overwrite=TRUE)
writeRaster(cn_f17_ndvi, filename="../outputs/ndvi/cn_f17_NDVI", format="GTiff", overwrite=TRUE)
writeRaster(cn_f20_ndvi, filename="../outputs/ndvi/cn_f20_NDVI", format="GTiff", overwrite=TRUE)
```

## Select bands to calculated NDVIre 
 (nir-red) / (nir+red)
 Band 1 - Green
 Band 2 = NIR
 Band 3 = Red 
 Band 4 = Red Edge

```{r}
alnus_f08_ndvire <- (alnus_f08_4band[[2]] - alnus_f08_4band[[4]]) / (alnus_f08_4band[[2]] + alnus_f08_4band[[4]])
cn_f17_ndvire = (cn_f17_4band[[2]] - cn_f17_4band[[4]]) / (cn_f17_4band[[2]] + cn_f17_4band[[4]])
cn_f20_ndvire = (cn_f20_4band[[2]] - cn_f20_4band[[4]]) / (cn_f20_4band[[2]] + cn_f20_4band[[4]])
```




```{r}
plot(alnus_f08_ndvire)
```

```{r}
plot(cn_f17_ndvire)
```

```{r}
plot(cn_f20_ndvire)
```

### Write to file
```{r eval=FALSE, include=FALSE}
writeRaster(alnus_f08_ndvire, filename="../outputs/ndvi/alnus_f08_NDVIre", format="GTiff", overwrite=TRUE)
writeRaster(cn_f17_ndvire, filename="../outputs/ndvi/cn_f17_NDVIre", format="GTiff", overwrite=TRUE)
writeRaster(cn_f20_ndvire, filename="../outputs/ndvi/cn_f20_NDVIre", format="GTiff", overwrite=TRUE)
```

## Select bands to calculated MSRre
 (nir-red) / (nir+red)
 Band 1 - Green
 Band 2 = NIR
 Band 3 = Red 
 Band 4 = Red Edge
```{r}
alnus_f08_msrre <- ((alnus_f08_4band[[2]] / alnus_f08_4band[[3]])-1 / sqrt((alnus_f08_4band[[2]] / alnus_f08_4band[[4]])+1))


cn_f17_ndvire = (cn_f17_4band[[2]] - cn_f17_4band[[4]]) / (cn_f17_4band[[2]] + cn_f17_4band[[4]])
cn_f20_ndvire = (cn_f20_4band[[2]] - cn_f20_4band[[4]]) / (cn_f20_4band[[2]] + cn_f20_4band[[4]])
```




```{r}
plot(alnus_f08_msrre )
```

```{r}
plot(cn_f17_ndvire)
```

```{r}
plot(cn_f20_ndvire)
```
