==> scripts/2.2_GCC-field-data.Rmd <==
---
title: "GCC and Field Data"
author: "Anna Talucci"
date: "2/24/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Overview

- Extract raster values based on field plot locations
- variable radius buffers applied
- Plot size for field data collection was variable from 4-10m radius plots
- apply 4m buffer

# Packages
```{r, include=FALSE}
library(sp) 
library(rgdal) 
library(raster) 
library(rgeos) # GIS Geometry
library(foreign) # deals with dbf files
library(dplyr)
library(ggplot2)
library(tidyr)
```


# RGB DATA



alnus_tr1_f07
ans_tr2_f04
ans_tr3_f06
brp_tr1_f09
brp_tr2_f12
cn_tr2_f17
cn_tr1_f15 


# GCC rasters

```{r}
alnus_tr1_f07 = raster("../data/gcc/alnus_tr1_f07_gcc.tif")
ans_tr2_f04 = raster("../data/gcc/ans_tr2_f04_gcc.tif")
ans_tr3_f06 = raster("../data/gcc/ans_tr3_f06_gcc.tif")
brp_tr1_f09 = raster("../data//gcc/brp_tr1_f09_gcc.tif")
brp_tr2_f12 = raster("../data/gcc/brp_tr2_f12_gcc.tif")
cn_tr2_f17 = raster("../data/gcc/cn_tr2_f17_gcc.tif")
cn_tr1_f15 = raster("../data/gcc/cn_tr1_f15_gcc.tif")
```


# Field points and buffers
hs2001 = readOGR("../data/hotspot_data/DL_FIRE_M6_2001/fire_archive_2001.shp", "fire_archive_2001")

## Point Data

Points
```{r}
alnus_tr1_pt = readOGR("../data/field-spatial/point/alnus_tr1_ee_pt.shp", "alnus_tr1_ee_pt")
ans_tr2_pt = readOGR("../data/field-spatial/point/ans_tr2_ee_pt.shp", "ans_tr2_ee_pt") 
ans_tr3_pt = readOGR("../data/field-spatial/point/ans_tr3_ee_pt.shp", "ans_tr3_ee_pt") 
brp_tr1_pt = readOGR("../data/field-spatial/point/bp_tr1_ee_pt.shp", "bp_tr1_ee_pt") 
brp_tr2_pt = readOGR("../data/field-spatial/point/bp_tr2_ee_pt.shp", "bp_tr2_ee_pt") 
cn_tr1_pt = readOGR("../data/field-spatial/point/cn_tr1_ee_pt.shp", "cn_tr1_ee_pt") 
cn_tr2_pt = readOGR("../data/field-spatial/point/cn_tr2_ee_pt.shp", "cn_tr2_ee_pt") 
```

```{r}
alnus_tr1_pt
```

```{r}
ans_tr2_pt
```

```{r}
ans_tr3_pt
```

```{r}
brp_tr1_pt
```

```{r}
brp_tr2_pt
```

```{r}
cn_tr1_pt
```

```{r}
cn_tr2_pt
```

```{r}
plot(cn_tr1_f15)
plot(cn_tr1_pt , add=TRUE)
```
## Function for functions
```{r}
my_summary <- function(x, na.rm) c(mean = mean(x, na.rm=na.rm), sd = sd(x, na.rm=na.rm))
```
min = min(x, na.rm=na.rm), max = max(x, na.rm=na.rm
## 25cm - extract mean pixel values
Use 25cm buffer for GCC as  function of distance from edge
extract(rasterA,points, buffer=150, fun=mean)

- Choose buffer based on pixel size and maintaining fine scale resolution but aggregating so that values are at the same spatial scale


```{r}
extract_alnus_tr1_25cm = raster::extract(alnus_tr1_f07, alnus_tr1_pt, buffer=.25, fun=mean, df=TRUE )
extract_ans_tr2_25cm = raster::extract(ans_tr2_f04, ans_tr2_pt, buffer=.25, fun=mean, df=TRUE )
extract_ans_tr3_25cm = raster::extract(ans_tr3_f06, ans_tr3_pt, buffer=.25, fun=mean,  df=TRUE )
extract_brp_tr1_25cm = raster::extract(brp_tr1_f09, brp_tr1_pt, buffer=.25, fun=mean, df=TRUE )
extract_brp_tr2_25cm = raster::extract(brp_tr2_f12, brp_tr2_pt, buffer=.25, fun=mean,  df=TRUE )
extract_cn_tr1_25cm = raster::extract(cn_tr1_f15, cn_tr1_pt, buffer=.25, fun=mean,  df=TRUE )
extract_cn_tr2_25cm = raster::extract(cn_tr2_f17, cn_tr2_pt, buffer=.25, fun=mean,  df=TRUE )
```
```{r}
sd_alnus_tr1_25cm = raster::extract(alnus_tr1_f07, alnus_tr1_pt, buffer=.25, fun=sd, df=TRUE )
sd_ans_tr2_25cm = raster::extract(ans_tr2_f04, ans_tr2_pt, buffer=.25, fun=sd, df=TRUE )
sd_ans_tr3_25cm = raster::extract(ans_tr3_f06, ans_tr3_pt, buffer=.25, fun=sd,  df=TRUE )
sd_brp_tr1_25cm = raster::extract(brp_tr1_f09, brp_tr1_pt, buffer=.25, fun=sd, df=TRUE )
sd_brp_tr2_25cm = raster::extract(brp_tr2_f12, brp_tr2_pt, buffer=.25, fun=sd,  df=TRUE )
sd_cn_tr1_25cm = raster::extract(cn_tr1_f15, cn_tr1_pt, buffer=.25, fun=sd,  df=TRUE )
sd_cn_tr2_25cm = raster::extract(cn_tr2_f17, cn_tr2_pt, buffer=.25, fun=sd,  df=TRUE )
```

```{r}
extract_alnus_tr1_25cm$plot_id = alnus_tr1_pt$ID
names(extract_alnus_tr1_25cm) = c('ID', 'mean_gcc_25cm', 'plot_id')
extract_alnus_tr1_25cm
```


```{r}
extract_ans_tr2_25cm$plot_id = ans_tr2_pt$ID
names(extract_ans_tr2_25cm) = c('ID', 'mean_gcc_25cm', 'plot_id')
extract_ans_tr2_25cm
```

```{r}
extract_ans_tr3_25cm$plot_id = ans_tr3_pt$ID
names(extract_ans_tr3_25cm) = c('ID', 'mean_gcc_25cm', 'plot_id')
extract_ans_tr3_25cm
```


```{r}
extract_brp_tr1_25cm$plot_id = brp_tr1_pt$ID
names(extract_brp_tr1_25cm) = c('ID', 'mean_gcc_25cm', 'plot_id')
extract_brp_tr1_25cm
```

```{r}
extract_brp_tr2_25cm$plot_id = brp_tr2_pt$ID
names(extract_brp_tr2_25cm) = c('ID', 'mean_gcc_25cm', 'plot_id')
extract_brp_tr2_25cm
```

```{r}
extract_cn_tr1_25cm$plot_id = cn_tr1_pt$ID
names(extract_cn_tr1_25cm) = c('ID', 'mean_gcc_25cm', 'plot_id')
extract_cn_tr1_25cm
```

```{r}
extract_cn_tr2_25cm$plot_id = cn_tr2_pt$ID
names(extract_cn_tr2_25cm) = c('ID', 'mean_gcc_25cm', 'plot_id')
extract_cn_tr2_25cm
```



```{r}
gcc_25cm = bind_rows(extract_alnus_tr1_25cm, extract_ans_tr2_25cm, extract_ans_tr3_25cm, extract_brp_tr1_25cm, extract_brp_tr2_25cm, extract_cn_tr1_25cm, extract_cn_tr2_25cm) %>% dplyr::select(plot_id, mean_gcc_25cm) %>% rename(ID = plot_id)
length(unique(gcc$ID))
gcc_25cm
```


```{r eval=FALSE, include=FALSE}
write.csv(gcc_25cm,"../outputs/extracted_data/gcc_25cm.csv", row.names = FALSE)
```

### SD
```{r}
sd_alnus_tr1_25cm$plot_id = alnus_tr1_pt$ID
names(sd_alnus_tr1_25cm) = c('ID', 'sd_gcc_25cm', 'plot_id')
sd_alnus_tr1_25cm
```


```{r}
sd_ans_tr2_25cm$plot_id = ans_tr2_pt$ID
names(sd_ans_tr2_25cm) = c('ID', 'sd_gcc_25cm', 'plot_id')
sd_ans_tr2_25cm
```

```{r}
sd_ans_tr3_25cm$plot_id = ans_tr3_pt$ID
names(sd_ans_tr3_25cm) = c('ID', 'sd_gcc_25cm', 'plot_id')
sd_ans_tr3_25cm
```


```{r}
sd_brp_tr1_25cm$plot_id = brp_tr1_pt$ID
names(sd_brp_tr1_25cm) = c('ID', 'sd_gcc_25cm', 'plot_id')
sd_brp_tr1_25cm
```

```{r}
sd_brp_tr2_25cm$plot_id = brp_tr2_pt$ID
names(sd_brp_tr2_25cm) = c('ID', 'sd_gcc_25cm', 'plot_id')
sd_brp_tr2_25cm
```

```{r}
sd_cn_tr1_25cm$plot_id = cn_tr1_pt$ID
names(sd_cn_tr1_25cm) = c('ID', 'sd_gcc_25cm', 'plot_id')
sd_cn_tr1_25cm
```

```{r}
sd_cn_tr2_25cm$plot_id = cn_tr2_pt$ID
names(sd_cn_tr2_25cm) = c('ID', 'sd_gcc_25cm', 'plot_id')
sd_cn_tr2_25cm
```



```{r}
gcc_25cm_sd = bind_rows(sd_alnus_tr1_25cm, sd_ans_tr2_25cm, sd_ans_tr3_25cm, sd_brp_tr1_25cm, sd_brp_tr2_25cm, sd_cn_tr1_25cm, sd_cn_tr2_25cm) %>% dplyr::select(plot_id, sd_gcc_25cm) %>% dplyr::rename(ID = plot_id)
length(unique(gcc$ID))
gcc_25cm_sd
```


```{r eval=FALSE, include=FALSE}
write.csv(gcc_25cm_sd,"../outputs/extracted_data/gcc_25cm_sd.csv", row.names = FALSE)
```

## 50cm - extract mean pixel values
Use 50cm buffer for GCC as  function of distance from edge
extract(rasterA,points, buffer=150, fun=mean)

- Choose buffer based on pixel size and maintaining fine scale resolution but aggregating so that values are at the same spatial scale

```{r}
extract_alnus_tr1_50cm = raster::extract(alnus_tr1_f07, alnus_tr1_pt, buffer=.50, fun=mean, df=TRUE )
extract_ans_tr2_50cm = raster::extract(ans_tr2_f04, ans_tr2_pt, buffer=.50, fun=mean, df=TRUE )
extract_ans_tr3_50cm = raster::extract(ans_tr3_f06, ans_tr3_pt, buffer=.50, fun=mean,  df=TRUE )
extract_brp_tr1_50cm = raster::extract(brp_tr1_f09, brp_tr1_pt, buffer=.50, fun=mean, df=TRUE )
extract_brp_tr2_50cm = raster::extract(brp_tr2_f12, brp_tr2_pt, buffer=.50, fun=mean,  df=TRUE )
extract_cn_tr1_50cm = raster::extract(cn_tr1_f15, cn_tr1_pt, buffer=.50, fun=mean,  df=TRUE )
extract_cn_tr2_50cm = raster::extract(cn_tr2_f17, cn_tr2_pt, buffer=.50, fun=mean,  df=TRUE )
```

```{r}
sd_alnus_tr1_50cm = raster::extract(alnus_tr1_f07, alnus_tr1_pt, buffer=.50, fun=sd, df=TRUE )
sd_ans_tr2_50cm = raster::extract(ans_tr2_f04, ans_tr2_pt, buffer=.50, fun=sd, df=TRUE )
sd_ans_tr3_50cm = raster::extract(ans_tr3_f06, ans_tr3_pt, buffer=.50, fun=sd,  df=TRUE )
sd_brp_tr1_50cm = raster::extract(brp_tr1_f09, brp_tr1_pt, buffer=.50, fun=sd, df=TRUE )
sd_brp_tr2_50cm = raster::extract(brp_tr2_f12, brp_tr2_pt, buffer=.50, fun=sd,  df=TRUE )
sd_cn_tr1_50cm = raster::extract(cn_tr1_f15, cn_tr1_pt, buffer=.50, fun=sd,  df=TRUE )
sd_cn_tr2_50cm = raster::extract(cn_tr2_f17, cn_tr2_pt, buffer=.50, fun=sd,  df=TRUE )
```

### Mean
```{r}
extract_alnus_tr1_50cm$plot_id = alnus_tr1_pt$ID
names(extract_alnus_tr1_50cm) = c('ID', 'mean_gcc_50cm', 'plot_id')
extract_alnus_tr1_50cm
```

```{r}
extract_ans_tr2_50cm$plot_id = ans_tr2_pt$ID
names(extract_ans_tr2_50cm) = c('ID', 'mean_gcc_50cm', 'plot_id')
extract_ans_tr2_50cm
```

```{r}
extract_ans_tr3_50cm$plot_id = ans_tr3_pt$ID
names(extract_ans_tr3_50cm) = c('ID', 'mean_gcc_50cm', 'plot_id')
extract_ans_tr3_50cm
```


```{r}
extract_brp_tr1_50cm$plot_id = brp_tr1_pt$ID
names(extract_brp_tr1_50cm) = c('ID', 'mean_gcc_50cm', 'plot_id')
extract_brp_tr1_50cm
```

```{r}
extract_brp_tr2_50cm$plot_id = brp_tr2_pt$ID
names(extract_brp_tr2_50cm) = c('ID', 'mean_gcc_50cm', 'plot_id')
extract_brp_tr2_50cm
```

```{r}
extract_cn_tr1_50cm$plot_id = cn_tr1_pt$ID
names(extract_cn_tr1_50cm) = c('ID', 'mean_gcc_50cm', 'plot_id')
extract_cn_tr1_50cm
```

```{r}
extract_cn_tr2_50cm$plot_id = cn_tr2_pt$ID
names(extract_cn_tr2_50cm) = c('ID', 'mean_gcc_50cm', 'plot_id')
extract_cn_tr2_50cm
```



```{r}
gcc_50cm = bind_rows(extract_alnus_tr1_50cm, extract_ans_tr2_50cm, extract_ans_tr3_50cm, extract_brp_tr1_50cm, extract_brp_tr2_50cm, extract_cn_tr1_50cm, extract_cn_tr2_50cm) %>% dplyr::select(plot_id, mean_gcc_50cm) %>% rename(ID = plot_id)
length(unique(gcc$ID))
gcc_50cm
```


```{r eval=FALSE, include=FALSE}
write.csv(gcc_50cm,"../outputs/extracted_data/gcc_50cm.csv", row.names = FALSE)
```

### SD

```{r}
sd_alnus_tr1_50cm$plot_id = alnus_tr1_pt$ID
names(sd_alnus_tr1_50cm) = c('ID', 'sd_gcc_50cm', 'plot_id')
sd_alnus_tr1_50cm
```

```{r}
sd_ans_tr2_50cm$plot_id = ans_tr2_pt$ID
names(sd_ans_tr2_50cm) = c('ID', 'sd_gcc_50cm', 'plot_id')
sd_ans_tr2_50cm
```

```{r}
sd_ans_tr3_50cm$plot_id = ans_tr3_pt$ID
names(sd_ans_tr3_50cm) = c('ID', 'sd_gcc_50cm', 'plot_id')
sd_ans_tr3_50cm
```


```{r}
sd_brp_tr1_50cm$plot_id = brp_tr1_pt$ID
names(sd_brp_tr1_50cm) = c('ID', 'sd_gcc_50cm', 'plot_id')
sd_brp_tr1_50cm
```

```{r}
sd_brp_tr2_50cm$plot_id = brp_tr2_pt$ID
names(sd_brp_tr2_50cm) = c('ID', 'sd_gcc_50cm', 'plot_id')
sd_brp_tr2_50cm
```

```{r}
sd_cn_tr1_50cm$plot_id = cn_tr1_pt$ID
names(sd_cn_tr1_50cm) = c('ID', 'sd_gcc_50cm', 'plot_id')
sd_cn_tr1_50cm
```

```{r}
sd_cn_tr2_50cm$plot_id = cn_tr2_pt$ID
names(sd_cn_tr2_50cm) = c('ID', 'sd_gcc_50cm', 'plot_id')
sd_cn_tr2_50cm
```



```{r}
gcc_50cm_sd = bind_rows(sd_alnus_tr1_50cm, sd_ans_tr2_50cm, sd_ans_tr3_50cm, sd_brp_tr1_50cm, sd_brp_tr2_50cm, sd_cn_tr1_50cm, sd_cn_tr2_50cm) %>% dplyr::select(plot_id, sd_gcc_50cm) %>% dplyr::rename(ID = plot_id)
length(unique(gcc$ID))
gcc_50cm_sd
```


```{r eval=FALSE, include=FALSE}
write.csv(gcc_50cm_sd,"../outputs/extracted_data/gcc_50cm_sd.csv", row.names = FALSE)
```


## 1m - extract mean pixel values
Use 1m buffer for GCC as  function of distance from edge
extract(rasterA,points, buffer=150, fun=mean)

- Choose buffer based on pixel size and maintaining fine scale resolution but aggregating so that values are at the same spatial scale
- Buffer is the radius of the buffer around each point

```{r}

extract_alnus_tr1 = raster::extract(alnus_tr1_f07, alnus_tr1_pt, buffer=1, fun=mean, df=TRUE )
extract_ans_tr2 = raster::extract(ans_tr2_f04, ans_tr2_pt, buffer=1, fun=mean, df=TRUE )
extract_ans_tr3 = raster::extract(ans_tr3_f06, ans_tr3_pt, buffer=1, fun=mean,  df=TRUE )
extract_brp_tr1 = raster::extract(brp_tr1_f09, brp_tr1_pt, buffer=1, fun=mean, df=TRUE )
extract_brp_tr2 = raster::extract(brp_tr2_f12, brp_tr2_pt, buffer=1, fun=mean,  df=TRUE )
extract_cn_tr1 = raster::extract(cn_tr1_f15, cn_tr1_pt, buffer=1, fun=mean,  df=TRUE )
extract_cn_tr2 = raster::extract(cn_tr2_f17, cn_tr2_pt, buffer=1, fun=mean,  df=TRUE )
```


```{r}

sd_alnus_tr1 = raster::extract(alnus_tr1_f07, alnus_tr1_pt, buffer=1, fun=sd, df=TRUE )
sd_ans_tr2 = raster::extract(ans_tr2_f04, ans_tr2_pt, buffer=1, fun=sd, df=TRUE )
sd_ans_tr3 = raster::extract(ans_tr3_f06, ans_tr3_pt, buffer=1, fun=sd,  df=TRUE )
sd_brp_tr1 = raster::extract(brp_tr1_f09, brp_tr1_pt, buffer=1, fun=sd, df=TRUE )
sd_brp_tr2 = raster::extract(brp_tr2_f12, brp_tr2_pt, buffer=1, fun=sd,  df=TRUE )
sd_cn_tr1 = raster::extract(cn_tr1_f15, cn_tr1_pt, buffer=1, fun=sd,  df=TRUE )
sd_cn_tr2 = raster::extract(cn_tr2_f17, cn_tr2_pt, buffer=1, fun=sd,  df=TRUE )
```

### Mean

```{r}
extract_alnus_tr1$plot_id = alnus_tr1_pt$ID
names(extract_alnus_tr1) = c('ID', 'mean_gcc_1m', 'plot_id')
extract_alnus_tr1
```

```{r}
extract_ans_tr2$plot_id = ans_tr2_pt$ID
names(extract_ans_tr2) = c('ID', 'mean_gcc_1m', 'plot_id')
extract_ans_tr2
```

```{r}
extract_ans_tr3$plot_id = ans_tr3_pt$ID
names(extract_ans_tr3) = c('ID', 'mean_gcc_1m', 'plot_id')
extract_ans_tr3
```


```{r}
extract_brp_tr1$plot_id = brp_tr1_pt$ID
names(extract_brp_tr1) = c('ID', 'mean_gcc_1m', 'plot_id')
extract_brp_tr1
```

```{r}
extract_brp_tr2$plot_id = brp_tr2_pt$ID
names(extract_brp_tr2) = c('ID', 'mean_gcc_1m', 'plot_id')
extract_brp_tr2
```

```{r}
extract_cn_tr1$plot_id = cn_tr1_pt$ID
names(extract_cn_tr1) = c('ID', 'mean_gcc_1m', 'plot_id')
extract_cn_tr1
```

```{r}
extract_cn_tr2$plot_id = cn_tr2_pt$ID
names(extract_cn_tr2) = c('ID', 'mean_gcc_1m', 'plot_id')
extract_cn_tr2
```

```{r}
gcc_1m = bind_rows(extract_alnus_tr1, extract_ans_tr2, extract_ans_tr3, extract_brp_tr1, extract_brp_tr2, extract_cn_tr1, extract_cn_tr2) %>% dplyr::select(plot_id, mean_gcc_1m) %>% rename(ID = plot_id)
length(unique(gcc_3m$ID))
gcc_1m
```

```{r eval=FALSE, include=FALSE}
write.csv(gcc_1m,"../outputs/extracted_data/gcc_1m.csv", row.names = FALSE)
```

### SD

```{r}
sd_alnus_tr1$plot_id = alnus_tr1_pt$ID
names(sd_alnus_tr1) = c('ID', 'sd_gcc_1m', 'plot_id')
sd_alnus_tr1
```

```{r}
sd_ans_tr2$plot_id = ans_tr2_pt$ID
names(sd_ans_tr2) = c('ID', 'sd_gcc_1m', 'plot_id')
sd_ans_tr2
```

```{r}
sd_ans_tr3$plot_id = ans_tr3_pt$ID
names(sd_ans_tr3) = c('ID', 'sd_gcc_1m', 'plot_id')
sd_ans_tr3
```


```{r}
sd_brp_tr1$plot_id = brp_tr1_pt$ID
names(sd_brp_tr1) = c('ID', 'sd_gcc_1m', 'plot_id')
sd_brp_tr1
```

```{r}
sd_brp_tr2$plot_id = brp_tr2_pt$ID
names(sd_brp_tr2) = c('ID', 'sd_gcc_1m', 'plot_id')
sd_brp_tr2
```

```{r}
sd_cn_tr1$plot_id = cn_tr1_pt$ID
names(sd_cn_tr1) = c('ID', 'sd_gcc_1m', 'plot_id')
sd_cn_tr1
```

```{r}
sd_cn_tr2$plot_id = cn_tr2_pt$ID
names(sd_cn_tr2) = c('ID', 'sd_gcc_1m', 'plot_id')
sd_cn_tr2
```

```{r}
gcc_1m_sd = bind_rows(sd_alnus_tr1, sd_ans_tr2, sd_ans_tr3, sd_brp_tr1, sd_brp_tr2, sd_cn_tr1, sd_cn_tr2) %>% dplyr::select(plot_id, sd_gcc_1m) %>% dplyr::rename(ID = plot_id)

gcc_1m_sd
```

```{r eval=FALSE, include=FALSE}
write.csv(gcc_1m,"../outputs/extracted_data/gcc_1m.csv", row.names = FALSE)
```

## 3m - extract mean pixel values
Extract a mean pixel value based on a 3m  radius for each plot point location. 
```{r}
extract_alnus_tr1_3m = raster::extract(alnus_tr1_f07, alnus_tr1_pt, buffer=3, fun=mean, df=TRUE )
extract_ans_tr2_3m = raster::extract(ans_tr2_f04, ans_tr2_pt, buffer=3, fun=mean, df=TRUE )
extract_ans_tr3_3m = raster::extract(ans_tr3_f06, ans_tr3_pt, buffer=3, fun=mean,  df=TRUE )
extract_brp_tr1_3m = raster::extract(brp_tr1_f09, brp_tr1_pt, buffer=3, fun=mean, df=TRUE )
extract_brp_tr2_3m = raster::extract(brp_tr2_f12, brp_tr2_pt, buffer=3, fun=mean,  df=TRUE )
extract_cn_tr1_3m = raster::extract(cn_tr1_f15, cn_tr1_pt, buffer=3, fun=mean,  df=TRUE )
extract_cn_tr2_3m = raster::extract(cn_tr2_f17, cn_tr2_pt, buffer=3, fun=mean,  df=TRUE )
```

```{r}
sd_alnus_tr1_3m = raster::extract(alnus_tr1_f07, alnus_tr1_pt, buffer=3, fun=sd, df=TRUE )
sd_ans_tr2_3m = raster::extract(ans_tr2_f04, ans_tr2_pt, buffer=3, fun=sd, df=TRUE )
sd_ans_tr3_3m = raster::extract(ans_tr3_f06, ans_tr3_pt, buffer=3, fun=sd,  df=TRUE )
sd_brp_tr1_3m = raster::extract(brp_tr1_f09, brp_tr1_pt, buffer=3, fun=sd, df=TRUE )
sd_brp_tr2_3m = raster::extract(brp_tr2_f12, brp_tr2_pt, buffer=3, fun=sd,  df=TRUE )
sd_cn_tr1_3m = raster::extract(cn_tr1_f15, cn_tr1_pt, buffer=3, fun=sd,  df=TRUE )
sd_cn_tr2_3m = raster::extract(cn_tr2_f17, cn_tr2_pt, buffer=3, fun=sd,  df=TRUE )
```
Rename columns for each transect so that the transects can be combine into one dataframe
### Mean
```{r}
extract_alnus_tr1_3m$plot_id = alnus_tr1_pt$ID
names(extract_alnus_tr1_3m) = c('ID', 'mean_gcc_3m', 'plot_id')
extract_alnus_tr1_3m
```

```{r}
extract_ans_tr2_3m$plot_id = ans_tr2_pt$ID
names(extract_ans_tr2_3m) = c('ID', 'mean_gcc_3m', 'plot_id')
extract_ans_tr2_3m
```

```{r}
extract_ans_tr3_3m$plot_id = ans_tr3_pt$ID
names(extract_ans_tr3_3m) = c('ID', 'mean_gcc_3m', 'plot_id')
extract_ans_tr3_3m
```


```{r}
extract_brp_tr1_3m$plot_id = brp_tr1_pt$ID
names(extract_brp_tr1_3m) = c('ID', 'mean_gcc_3m', 'plot_id')
extract_brp_tr1_3m
```

```{r}
extract_brp_tr2_3m$plot_id = brp_tr2_pt$ID
names(extract_brp_tr2_3m) = c('ID', 'mean_gcc_3m', 'plot_id')
extract_brp_tr2_3m
```

```{r}
extract_cn_tr1_3m$plot_id = cn_tr1_pt$ID
names(extract_cn_tr1_3m) = c('ID', 'mean_gcc_3m', 'plot_id')
extract_cn_tr1_3m
```

```{r}
extract_cn_tr2_3m$plot_id = cn_tr2_pt$ID
names(extract_cn_tr2_3m) = c('ID', 'mean_gcc_3m', 'plot_id')
extract_cn_tr2_3m
```

Combine transect data into a single dataframe
```{r}
gcc_3m = bind_rows(extract_alnus_tr1_3m, extract_ans_tr2_3m, extract_ans_tr3_3m, extract_brp_tr1_3m, extract_brp_tr2_3m, extract_cn_tr1_3m, extract_cn_tr2_3m) %>% dplyr::select(plot_id, mean_gcc_3m) %>% rename(ID = plot_id)
length(unique(gcc_3m$ID))
gcc_3m
```

Save datframe to a csv file
```{r eval=FALSE, include=FALSE}
write.csv(gcc_3m,"../outputs/extracted_data/gcc_3m.csv", row.names = FALSE)
```

### SD
```{r}
sd_alnus_tr1_3m$plot_id = alnus_tr1_pt$ID
names(sd_alnus_tr1_3m) = c('ID', 'sd_gcc_3m', 'plot_id')
sd_alnus_tr1_3m
```

```{r}
sd_ans_tr2_3m$plot_id = ans_tr2_pt$ID
names(sd_ans_tr2_3m) = c('ID', 'sd_gcc_3m', 'plot_id')
sd_ans_tr2_3m
```

```{r}
sd_ans_tr3_3m$plot_id = ans_tr3_pt$ID
names(sd_ans_tr3_3m) = c('ID', 'sd_gcc_3m', 'plot_id')
sd_ans_tr3_3m
```


```{r}
sd_brp_tr1_3m$plot_id = brp_tr1_pt$ID
names(sd_brp_tr1_3m) = c('ID', 'sd_gcc_3m', 'plot_id')
sd_brp_tr1_3m
```

```{r}
sd_brp_tr2_3m$plot_id = brp_tr2_pt$ID
names(sd_brp_tr2_3m) = c('ID', 'sd_gcc_3m', 'plot_id')
sd_brp_tr2_3m
```

```{r}
sd_cn_tr1_3m$plot_id = cn_tr1_pt$ID
names(sd_cn_tr1_3m) = c('ID', 'sd_gcc_3m', 'plot_id')
sd_cn_tr1_3m
```

```{r}
sd_cn_tr2_3m$plot_id = cn_tr2_pt$ID
names(sd_cn_tr2_3m) = c('ID', 'sd_gcc_3m', 'plot_id')
sd_cn_tr2_3m
```

Combine transect data into a single dataframe
```{r}
gcc_3m_sd = bind_rows(sd_alnus_tr1_3m, sd_ans_tr2_3m, sd_ans_tr3_3m, sd_brp_tr1_3m, sd_brp_tr2_3m, sd_cn_tr1_3m, sd_cn_tr2_3m) %>% dplyr::select(plot_id, sd_gcc_3m) %>% dplyr::rename(ID = plot_id)

gcc_3m_sd
```

Save datframe to a csv file
```{r eval=FALSE, include=FALSE}
write.csv(gcc_3m_sd,"../outputs/extracted_data/gcc_3m_sd.csv", row.names = FALSE)
```


## 5m - extract mean pixel values
Extract a mean pixel value based on a 5m  radius for each plot point location. 
```{r}
extract_alnus_tr1_5m = raster::extract(alnus_tr1_f07, alnus_tr1_pt, buffer=5, fun=mean, df=TRUE )
extract_ans_tr2_5m = raster::extract(ans_tr2_f04, ans_tr2_pt, buffer=5, fun=mean, df=TRUE )
extract_ans_tr3_5m = raster::extract(ans_tr3_f06, ans_tr3_pt, buffer=5, fun=mean,  df=TRUE )
extract_brp_tr1_5m = raster::extract(brp_tr1_f09, brp_tr1_pt, buffer=5, fun=mean, df=TRUE )
extract_brp_tr2_5m = raster::extract(brp_tr2_f12, brp_tr2_pt, buffer=5, fun=mean,  df=TRUE )
extract_cn_tr1_5m = raster::extract(cn_tr1_f15, cn_tr1_pt, buffer=5, fun=mean,  df=TRUE )
extract_cn_tr2_5m = raster::extract(cn_tr2_f17, cn_tr2_pt, buffer=5, fun=mean,  df=TRUE )
```

```{r}
sd_alnus_tr1_5m = raster::extract(alnus_tr1_f07, alnus_tr1_pt, buffer=5, fun=sd, df=TRUE )
sd_ans_tr2_5m = raster::extract(ans_tr2_f04, ans_tr2_pt, buffer=5, fun=sd, df=TRUE )
sd_ans_tr3_5m = raster::extract(ans_tr3_f06, ans_tr3_pt, buffer=5, fun=sd,  df=TRUE )
sd_brp_tr1_5m = raster::extract(brp_tr1_f09, brp_tr1_pt, buffer=5, fun=sd, df=TRUE )
sd_brp_tr2_5m = raster::extract(brp_tr2_f12, brp_tr2_pt, buffer=5, fun=sd,  df=TRUE )
sd_cn_tr1_5m = raster::extract(cn_tr1_f15, cn_tr1_pt, buffer=5, fun=sd,  df=TRUE )
sd_cn_tr2_5m = raster::extract(cn_tr2_f17, cn_tr2_pt, buffer=5, fun=sd,  df=TRUE )
```

### Mean
Rename columns for each transect so that the transects can be combine into one dataframe

```{r}
extract_alnus_tr1_5m$plot_id = alnus_tr1_pt$ID
names(extract_alnus_tr1_5m) = c('ID', 'mean_gcc_5m', 'plot_id')
extract_alnus_tr1_5m
```

```{r}
extract_ans_tr2_5m$plot_id = ans_tr2_pt$ID
names(extract_ans_tr2_5m) = c('ID', 'mean_gcc_5m', 'plot_id')
extract_ans_tr2_5m
```

```{r}
extract_ans_tr3_5m$plot_id = ans_tr3_pt$ID
names(extract_ans_tr3_5m) = c('ID', 'mean_gcc_5m', 'plot_id')
extract_ans_tr3_5m
```


```{r}
extract_brp_tr1_5m$plot_id = brp_tr1_pt$ID
names(extract_brp_tr1_5m) = c('ID', 'mean_gcc_5m', 'plot_id')
extract_brp_tr1_5m
```

```{r}
extract_brp_tr2_5m$plot_id = brp_tr2_pt$ID
names(extract_brp_tr2_5m) = c('ID', 'mean_gcc_5m', 'plot_id')
extract_brp_tr2_5m
```

```{r}
extract_cn_tr1_5m$plot_id = cn_tr1_pt$ID
names(extract_cn_tr1_5m) = c('ID', 'mean_gcc_5m', 'plot_id')
extract_cn_tr1_5m
```

```{r}
extract_cn_tr2_5m$plot_id = cn_tr2_pt$ID
names(extract_cn_tr2_5m) = c('ID', 'mean_gcc_5m', 'plot_id')
extract_cn_tr2_5m
```

Combine transect data into a single dataframe
```{r}
gcc_5m = bind_rows(extract_alnus_tr1_5m, extract_ans_tr2_5m, extract_ans_tr3_5m, extract_brp_tr1_5m, extract_brp_tr2_5m, extract_cn_tr1_5m, extract_cn_tr2_5m) %>% dplyr::select(plot_id, mean_gcc_5m) %>% rename(ID = plot_id)
length(unique(gcc_5m$ID))
gcc_5m
```

Save dataframe to a csv file
```{r eval=FALSE, include=FALSE}
write.csv(gcc_5m,"../outputs/extracted_data/gcc_5m.csv", row.names = FALSE)
```

### SD
Rename columns for each transect so that the transects can be combine into one dataframe

```{r}
sd_alnus_tr1_5m$plot_id = alnus_tr1_pt$ID
names(sd_alnus_tr1_5m) = c('ID', 'sd_gcc_5m', 'plot_id')
sd_alnus_tr1_5m
```

```{r}
sd_ans_tr2_5m$plot_id = ans_tr2_pt$ID
names(sd_ans_tr2_5m) = c('ID', 'sd_gcc_5m', 'plot_id')
sd_ans_tr2_5m
```

```{r}
sd_ans_tr3_5m$plot_id = ans_tr3_pt$ID
names(sd_ans_tr3_5m) = c('ID', 'sd_gcc_5m', 'plot_id')
sd_ans_tr3_5m
```


```{r}
sd_brp_tr1_5m$plot_id = brp_tr1_pt$ID
names(sd_brp_tr1_5m) = c('ID', 'sd_gcc_5m', 'plot_id')
sd_brp_tr1_5m
```

```{r}
sd_brp_tr2_5m$plot_id = brp_tr2_pt$ID
names(sd_brp_tr2_5m) = c('ID', 'sd_gcc_5m', 'plot_id')
sd_brp_tr2_5m
```

```{r}
sd_cn_tr1_5m$plot_id = cn_tr1_pt$ID
names(sd_cn_tr1_5m) = c('ID', 'sd_gcc_5m', 'plot_id')
sd_cn_tr1_5m
```

```{r}
sd_cn_tr2_5m$plot_id = cn_tr2_pt$ID
names(sd_cn_tr2_5m) = c('ID', 'sd_gcc_5m', 'plot_id')
sd_cn_tr2_5m
```

Combine transect data into a single dataframe
```{r}
gcc_5m_sd = bind_rows(sd_alnus_tr1_5m, sd_ans_tr2_5m, sd_ans_tr3_5m, sd_brp_tr1_5m, sd_brp_tr2_5m, sd_cn_tr1_5m, sd_cn_tr2_5m) %>% dplyr::select(plot_id, sd_gcc_5m) %>% dplyr::rename(ID = plot_id)

gcc_5m_sd
```

Save dataframe to a csv file
```{r eval=FALSE, include=FALSE}
write.csv(gcc_5m,"../outputs/extracted_data/gcc_5m.csv", row.names = FALSE)
```


## 10m - extract mean pixel values
Extract a mean pixel value based on a 10m  radius for each plot point location. 
```{r}
extract_alnus_tr1_10m = raster::extract(alnus_tr1_f07, alnus_tr1_pt, buffer=10, fun=mean, df=TRUE )
extract_ans_tr2_10m = raster::extract(ans_tr2_f04, ans_tr2_pt, buffer=10, fun=mean, df=TRUE )
extract_ans_tr3_10m = raster::extract(ans_tr3_f06, ans_tr3_pt, buffer=10, fun=mean,  df=TRUE )
extract_brp_tr1_10m = raster::extract(brp_tr1_f09, brp_tr1_pt, buffer=10, fun=mean, df=TRUE )
extract_brp_tr2_10m = raster::extract(brp_tr2_f12, brp_tr2_pt, buffer=10, fun=mean,  df=TRUE )
extract_cn_tr1_10m = raster::extract(cn_tr1_f15, cn_tr1_pt, buffer=10, fun=mean,  df=TRUE )
extract_cn_tr2_10m = raster::extract(cn_tr2_f17, cn_tr2_pt, buffer=10, fun=mean,  df=TRUE )
```

```{r}
sd_alnus_tr1_10m = raster::extract(alnus_tr1_f07, alnus_tr1_pt, buffer=10, fun=sd, df=TRUE )
sd_ans_tr2_10m = raster::extract(ans_tr2_f04, ans_tr2_pt, buffer=10, fun=sd, df=TRUE )
sd_ans_tr3_10m = raster::extract(ans_tr3_f06, ans_tr3_pt, buffer=10, fun=sd,  df=TRUE )
sd_brp_tr1_10m = raster::extract(brp_tr1_f09, brp_tr1_pt, buffer=10, fun=sd, df=TRUE )
sd_brp_tr2_10m = raster::extract(brp_tr2_f12, brp_tr2_pt, buffer=10, fun=sd,  df=TRUE )
sd_cn_tr1_10m = raster::extract(cn_tr1_f15, cn_tr1_pt, buffer=10, fun=sd,  df=TRUE )
sd_cn_tr2_10m = raster::extract(cn_tr2_f17, cn_tr2_pt, buffer=10, fun=sd,  df=TRUE )
```

###Mean
Rename columns for each transect so that the transects can be combine into one dataframe
```{r}
extract_alnus_tr1_10m$plot_id = alnus_tr1_pt$ID
names(extract_alnus_tr1_10m) = c('ID', 'mean_gcc_10m', 'plot_id')
extract_alnus_tr1_10m
```

```{r}
extract_ans_tr2_10m$plot_id = ans_tr2_pt$ID
names(extract_ans_tr2_10m) = c('ID', 'mean_gcc_10m', 'plot_id')
extract_ans_tr2_10m
```

```{r}
extract_ans_tr3_10m$plot_id = ans_tr3_pt$ID
names(extract_ans_tr3_10m) = c('ID', 'mean_gcc_10m', 'plot_id')
extract_ans_tr3_10m
```


```{r}
extract_brp_tr1_10m$plot_id = brp_tr1_pt$ID
names(extract_brp_tr1_10m) = c('ID', 'mean_gcc_10m', 'plot_id')
extract_brp_tr1_10m
```

```{r}
extract_brp_tr2_10m$plot_id = brp_tr2_pt$ID
names(extract_brp_tr2_10m) = c('ID', 'mean_gcc_10m', 'plot_id')
extract_brp_tr2_10m
```

```{r}
extract_cn_tr1_10m$plot_id = cn_tr1_pt$ID
names(extract_cn_tr1_10m) = c('ID', 'mean_gcc_10m', 'plot_id')
extract_cn_tr1_10m
```

```{r}
extract_cn_tr2_10m$plot_id = cn_tr2_pt$ID
names(extract_cn_tr2_10m) = c('ID', 'mean_gcc_10m', 'plot_id')
extract_cn_tr2_10m
```

Combine transect data into a single dataframe
```{r}
gcc_10m = bind_rows(extract_alnus_tr1_10m, extract_ans_tr2_10m, extract_ans_tr3_10m, extract_brp_tr1_10m, extract_brp_tr2_10m, extract_cn_tr1_10m, extract_cn_tr2_10m) %>% dplyr::select(plot_id, mean_gcc_10m) %>% rename(ID = plot_id)
length(unique(gcc_10m$ID))
gcc_10m
```

Save datframe to a csv file
```{r eval=FALSE, include=FALSE}
write.csv(gcc_10m,"../outputs/extracted_data/gcc_10m.csv", row.names = FALSE)
```


###SD
Rename columns for each transect so that the transects can be combine into one dataframe
```{r}
sd_alnus_tr1_10m$plot_id = alnus_tr1_pt$ID
names(sd_alnus_tr1_10m) = c('ID', 'sd_gcc_10m', 'plot_id')
sd_alnus_tr1_10m
```

```{r}
sd_ans_tr2_10m$plot_id = ans_tr2_pt$ID
names(sd_ans_tr2_10m) = c('ID', 'sd_gcc_10m', 'plot_id')
sd_ans_tr2_10m
```

```{r}
sd_ans_tr3_10m$plot_id = ans_tr3_pt$ID
names(sd_ans_tr3_10m) = c('ID', 'sd_gcc_10m', 'plot_id')
sd_ans_tr3_10m
```


```{r}
sd_brp_tr1_10m$plot_id = brp_tr1_pt$ID
names(sd_brp_tr1_10m) = c('ID', 'sd_gcc_10m', 'plot_id')
sd_brp_tr1_10m
```

```{r}
sd_brp_tr2_10m$plot_id = brp_tr2_pt$ID
names(sd_brp_tr2_10m) = c('ID', 'sd_gcc_10m', 'plot_id')
sd_brp_tr2_10m
```

```{r}
sd_cn_tr1_10m$plot_id = cn_tr1_pt$ID
names(sd_cn_tr1_10m) = c('ID', 'sd_gcc_10m', 'plot_id')
sd_cn_tr1_10m
```

```{r}
sd_cn_tr2_10m$plot_id = cn_tr2_pt$ID
names(sd_cn_tr2_10m) = c('ID', 'sd_gcc_10m', 'plot_id')
sd_cn_tr2_10m
```

Combine transect data into a single dataframe
```{r}
gcc_10m_sd = bind_rows(sd_alnus_tr1_10m, sd_ans_tr2_10m, sd_ans_tr3_10m, sd_brp_tr1_10m, sd_brp_tr2_10m, sd_cn_tr1_10m, sd_cn_tr2_10m) %>% dplyr::select(plot_id, sd_gcc_10m) %>% dplyr::rename(ID = plot_id)

gcc_10m_sd
```

Save datframe to a csv file
```{r eval=FALSE, include=FALSE}
write.csv(gcc_10m_sd,"../outputs/extracted_data/gcc_10m_sd.csv", row.names = FALSE)
```


## Trees -variable plot size extract mean pixel values
Use field plot size, which varied by plot.

```{r}
names(alnus_tr1_pt)
```

```{r}

extract_alnus_tr1_tree = raster::extract(alnus_tr1_f07, alnus_tr1_pt, buffer=alnus_tr1_pt$radius, fun=mean, df=TRUE )
extract_ans_tr2_tree = raster::extract(ans_tr2_f04, ans_tr2_pt, buffer=ans_tr2_pt$radius, fun=mean, df=TRUE )
extract_ans_tr3_tree = raster::extract(ans_tr3_f06, ans_tr3_pt, buffer=ans_tr3_pt$radius, fun=mean,  df=TRUE )
extract_brp_tr1_tree = raster::extract(brp_tr1_f09, brp_tr1_pt, buffer=brp_tr1_pt$radius, fun=mean, df=TRUE )
extract_brp_tr2_tree = raster::extract(brp_tr2_f12, brp_tr2_pt, buffer=brp_tr2_pt$radius, fun=mean,  df=TRUE )
extract_cn_tr1_tree = raster::extract(cn_tr1_f15, cn_tr1_pt, buffer=cn_tr1_pt$radius, fun=mean,  df=TRUE )
extract_cn_tr2_tree = raster::extract(cn_tr2_f17, cn_tr2_pt, buffer=cn_tr2_pt$radius, fun=mean,  df=TRUE )
```


```{r}
extract_alnus_tr1_tree$plot_id = alnus_tr1_pt$ID
names(extract_alnus_tr1_tree) = c('ID', 'mean_gcc_tree', 'plot_id')
extract_alnus_tr1_tree
```

```{r}
extract_ans_tr2_tree$plot_id = ans_tr2_pt$ID
names(extract_ans_tr2_tree) = c('ID', 'mean_gcc_tree', 'plot_id')
extract_ans_tr2_tree
```

```{r}
extract_ans_tr3_tree$plot_id = ans_tr3_pt$ID
names(extract_ans_tr3_tree) = c('ID', 'mean_gcc_tree', 'plot_id')
extract_ans_tr3_tree
```


```{r}
extract_brp_tr1_tree$plot_id = brp_tr1_pt$ID
names(extract_brp_tr1_tree) = c('ID', 'mean_gcc_tree', 'plot_id')
extract_brp_tr1_tree
```

```{r}
extract_brp_tr2_tree$plot_id = brp_tr2_pt$ID
names(extract_brp_tr2_tree) = c('ID', 'mean_gcc_tree', 'plot_id')
extract_brp_tr2_tree
```

```{r}
extract_cn_tr1_tree$plot_id = cn_tr1_pt$ID
names(extract_cn_tr1_tree) = c('ID', 'mean_gcc_tree', 'plot_id')
extract_cn_tr1_tree
```

```{r}
extract_cn_tr2_tree$plot_id = cn_tr2_pt$ID
names(extract_cn_tr2_tree) = c('ID', 'mean_gcc_tree', 'plot_id')
extract_cn_tr2_tree
```


```{r}
gcc_tree = bind_rows(extract_alnus_tr1_tree, extract_ans_tr2_tree, extract_ans_tr3_tree, extract_brp_tr1_tree, extract_brp_tr2_tree, extract_cn_tr1_tree, extract_cn_tr2_tree) %>% dplyr::select(plot_id, mean_gcc_tree) %>% rename(ID = plot_id)
length(unique(gcc_tree$ID))
gcc_tree
```


```{r eval=FALSE, include=FALSE}
write.csv(gcc_tree,"../outputs/extracted_data/gcc_tree.csv", row.names = FALSE)
```

## Shrub -variable plot size extract mean pixel values
```{r}
head(ans_tr2_pt)
```

```{r}
extract_alnus_tr1_shrub = raster::extract(alnus_tr1_f07, alnus_tr1_pt, buffer=alnus_tr1_pt$plt_lng, fun=mean, df=TRUE )
extract_ans_tr2_shrub = raster::extract(ans_tr2_f04, ans_tr2_pt, buffer=ans_tr2_pt$plt_lng, fun=mean, df=TRUE )
extract_ans_tr3_shrub = raster::extract(ans_tr3_f06, ans_tr3_pt, buffer=ans_tr3_pt$plt_lng, fun=mean,  df=TRUE )
extract_brp_tr1_shrub = raster::extract(brp_tr1_f09, brp_tr1_pt, buffer=brp_tr1_pt$plt_lng, fun=mean, df=TRUE )
extract_brp_tr2_shrub = raster::extract(brp_tr2_f12, brp_tr2_pt, buffer=brp_tr2_pt$plt_lng, fun=mean,  df=TRUE )
extract_cn_tr1_shrub = raster::extract(cn_tr1_f15, cn_tr1_pt, buffer=cn_tr1_pt$plt_lng, fun=mean,  df=TRUE )
extract_cn_tr2_shrub = raster::extract(cn_tr2_f17, cn_tr2_pt, buffer=cn_tr2_pt$plt_lng, fun=mean,  df=TRUE )
```


```{r}
extract_alnus_tr1_shrub$plot_id = alnus_tr1_pt$ID
names(extract_alnus_tr1_shrub) = c('ID', 'mean_gcc_shrub', 'plot_id')
extract_alnus_tr1_shrub
```

```{r}
extract_ans_tr2_shrub$plot_id = ans_tr2_pt$ID
names(extract_ans_tr2_shrub) = c('ID', 'mean_gcc_shrub', 'plot_id')
extract_ans_tr2_shrub
```

```{r}
extract_ans_tr3_shrub$plot_id = ans_tr3_pt$ID
names(extract_ans_tr3_shrub) = c('ID', 'mean_gcc_shrub', 'plot_id')
extract_ans_tr3_shrub
```


```{r}
extract_brp_tr1_shrub$plot_id = brp_tr1_pt$ID
names(extract_brp_tr1_shrub) = c('ID', 'mean_gcc_shrub', 'plot_id')
extract_brp_tr1_shrub
```

```{r}
extract_brp_tr2_shrub$plot_id = brp_tr2_pt$ID
names(extract_brp_tr2_shrub) = c('ID', 'mean_gcc_shrub', 'plot_id')
extract_brp_tr2_shrub
```

```{r}
extract_cn_tr1_shrub$plot_id = cn_tr1_pt$ID
names(extract_cn_tr1_shrub) = c('ID', 'mean_gcc_shrub', 'plot_id')
extract_cn_tr1_shrub
```

```{r}
extract_cn_tr2_shrub$plot_id = cn_tr2_pt$ID
names(extract_cn_tr2_shrub) = c('ID', 'mean_gcc_shrub', 'plot_id')
extract_cn_tr2_shrub
```

```{r}
gcc_shrub = bind_rows(extract_alnus_tr1_shrub, extract_ans_tr2_shrub, extract_ans_tr3_shrub, extract_brp_tr1_shrub, extract_brp_tr2_shrub, extract_cn_tr1_shrub, extract_cn_tr2_shrub) %>% dplyr::select(plot_id, mean_gcc_shrub) %>% rename(ID = plot_id)
length(unique(gcc_shrub$ID))
gcc_shrub
```


```{r eval=FALSE, include=FALSE}
write.csv(gcc_shrub,"../outputs/extracted_data/gcc_shrub.csv", row.names = FALSE)
```


# Combine gcc data

##Mean
```{r}
gcc_merge1 = merge(gcc_25cm, gcc_50cm,  by= "ID")
gcc_merge2 = merge(gcc_1m, gcc_3m,  by= "ID")
gcc_merge3 = merge(gcc_5m, gcc_10m,  by= "ID")
gcc_merge4 = merge(gcc_tree, gcc_shrub, by="ID")
gcc_merge5 = merge(gcc_merge1, gcc_merge2,  by="ID")
gcc_merge6 = merge(gcc_merge3, gcc_merge4,  by="ID")
gcc_all = merge(gcc_merge5, gcc_merge6, by="ID")
gcc_all
```


```{r}
write.csv(gcc_all,"../outputs/extracted_data/gcc_all.csv", row.names = FALSE)
```

### Sd
```{r}
gcc_merge1_sd = merge(gcc_25cm_sd, gcc_50cm_sd,  by= "ID")
gcc_merge2_sd = merge(gcc_1m_sd, gcc_3m_sd,  by= "ID")
gcc_merge3_sd = merge(gcc_5m_sd, gcc_10m_sd,  by= "ID")
gcc_merge5_sd = merge(gcc_merge1_sd, gcc_merge2_sd,  by="ID")
gcc_all_sd = merge(gcc_merge5_sd, gcc_merge3_sd,  by="ID")
gcc_all_sd
```


```{r}
write.csv(gcc_all_sd,"../outputs/extracted_data/gcc_all_sd.csv", row.names = FALSE)
```

# Extract all pixel values in 10m
## 10m - extract all pixel values
Extract a mean pixel value based on a 10m  radius for each plot point location. 
```{r}
extract_alnus_tr1_10m_all = raster::extract(alnus_tr1_f07, alnus_tr1_pt, buffer=10, df=TRUE )

extract_brp_tr2_10m_all = raster::extract(brp_tr2_f12, brp_tr2_pt, buffer=10,  df=TRUE )
extract_cn_tr1_10m_all = raster::extract(cn_tr1_f15, cn_tr1_pt, buffer=10,  df=TRUE )
extract_cn_tr2_10m_all = raster::extract(cn_tr2_f17, cn_tr2_pt, buffer=10,  df=TRUE )
```

### alnus add plot id
```{r}
alnus_tr1_pt@data
```
```{r}
df_extract_alnus_tr1_10m_all = data.frame(extract_alnus_tr1_10m_all)
```

```{r}
head(df_extract_alnus_tr1_10m_all)
tail(df_extract_alnus_tr1_10m_all)
```


```{r}
alnus_gcc = df_extract_alnus_tr1_10m_all %>%
  mutate(plot_id = ifelse(ID == "1", "Alnus_1984_1_-25",
                    ifelse(ID == "2", "Alnus_1984_1_-50",
                    ifelse(ID == "3", "Alnus_1984_1_100",
                    ifelse(ID == "4", "Alnus_1984_1_150", 
                    ifelse(ID == "5", "Alnus_1984_1_200",
                    ifelse(ID == "6", "Alnus_1984_1_25",
                    ifelse(ID == "7", "Alnus_1984_1_50", "Alnus_1984_1_75")))))))) %>%
  dplyr::rename(gcc = alnus_tr1_f07_gcc)
```

```{r}
head(alnus_gcc)
```



### bp tr2 add plot id
```{r}
brp_tr2_pt@data
```
```{r}
df_extract_bp_tr2_10m_all = data.frame(extract_brp_tr2_10m_all)
head(extract_brp_tr2_10m_all)
```


```{r}
bptr2_gcc = df_extract_bp_tr2_10m_all %>%
  mutate(plot_id = ifelse(ID == "1", "BP_1983_2_-25a",
                    ifelse(ID == "2", "BP_1983_2_-25b",
                    ifelse(ID == "3", "BP_1983_2_100",
                    ifelse(ID == "4", "BP_1983_2_150", 
                    ifelse(ID == "5", "BP_1983_2_200",
                    ifelse(ID == "6", "BP_1983_2_25",
                    ifelse(ID == "7", "BP_1983_2_50", "BP_1983_2_75")))))))) %>%
  dplyr::rename(gcc = brp_tr2_f12_gcc)

head(bptr2_gcc)
```



### CN tr1 add plot id
```{r}
cn_tr1_pt@data
```
```{r}
df_extract_cn_tr1_10m_all = data.frame(extract_cn_tr1_10m_all)
head(df_extract_cn_tr1_10m_all)
```


```{r}
cntr1_gcc = df_extract_cn_tr1_10m_all %>%
  mutate(plot_id = ifelse(ID == "1", "CN_2001_1_-25",
                    ifelse(ID == "2", "CN_2001_1_-50",
                    ifelse(ID == "3", "CN_2001_1_100",
                    ifelse(ID == "4", "CN_2001_1_150", 
                    ifelse(ID == "5", "CN_2001_1_200",
                    ifelse(ID == "6", "CN_2001_1_25",
                    ifelse(ID == "7", "CN_2001_1_50", "CN_2001_1_75")))))))) %>%
  dplyr::rename(gcc = cn_tr1_f15_gcc)

head(cntr1_gcc)
```



### CN tr2 add plot id
```{r}
cn_tr2_pt@data
```
```{r}
df_extract_cn_tr2_10m_all = data.frame(extract_cn_tr2_10m_all)
head(df_extract_cn_tr2_10m_all)
```


```{r}
cntr2_gcc = df_extract_cn_tr2_10m_all %>%
  mutate(plot_id = ifelse(ID == "1", "CN_2001_2_-35",
                    ifelse(ID == "2", "CN_2001_2_-60",
                    ifelse(ID == "3", "CN_2001_2_100",
                    ifelse(ID == "4", "CN_2001_2_150", 
                    ifelse(ID == "5", "CN_2001_2_200",
                    ifelse(ID == "6", "CN_2001_2_25",
                    ifelse(ID == "7", "CN_2001_2_50", "CN_2001_2_75")))))))) %>%
  dplyr::rename(gcc = cn_tr2_f17_gcc)

head(cntr2_gcc)
```

## Combine in to single data frame

```{r}
gcc_10m_all =  bind_rows(alnus_gcc, bptr2_gcc, cntr1_gcc, cntr2_gcc) 
summary(gcc_10m_all)
```

```{r}
write.csv(gcc_10m_all,"../outputs/extracted_data/2020-08-17_gcc_all_pixels.csv", row.names = FALSE)
```
