==> scripts/2.3_NDVI-field-extract.Rmd <==
---
title: "NDVI-field-extract"
author: "Anna Talucci"
date: "2/28/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Overview

- Extract mean NDVI values with buffers



- Plot size for field data collection was variable from 4-10m radius plots
- apply 4m buffer

# Packages
```{r, include=FALSE}
library(sp) 
library(rgdal) 
library(raster) 
library(rgeos) # GIS Geometry
library(foreign) # deals with dbf files
library(dplyr)
library(ggplot2)
```


# NVDI DATA

- alnus_tr1_f08
- bp_tr2_f1
- cn_tr2_f17
- cn_tr1_f20 

# NDVI rasters (Pix4D calcualted)

```{r}
alnus_tr1_f08_ndvi = raster("../data/pix4d_ndvi_Aug10/RU_ALN_TR1_FL008M_ndvi.tif")
bp_tr2_f13_ndvi = raster("../data/pix4d_ndvi_Aug10/RU_BP_TR2_FL013M_ndvi.tif")
cn_tr2_f17_ndvi = raster("../data/pix4d_ndvi_Aug10/RU_CYN_TR2_FL017M_ndvi.tif")
cn_tr1_f20_ndvi = raster("../data/pix4d_ndvi_Aug10/RU_CYN_TR1_FL020M_ndvi.tif")
```

view raster meta data
```{r}
alnus_tr1_f08_ndvi
bp_tr2_f13_ndvi
cn_tr2_f17_ndvi
cn_tr1_f20_ndvi
```


# Field point data

```{r}
alnus_tr1_pt = readOGR("../data/field-spatial/point/alnus_tr1_ee_pt.shp", "alnus_tr1_ee_pt")
bp_tr2_pt = readOGR("../data/field-spatial/point/bp_tr2_ee_pt.shp", "bp_tr2_ee_pt") 
cn_tr1_pt = readOGR("../data/field-spatial/point/cn_tr1_ee_pt.shp", "cn_tr1_ee_pt") 
cn_tr2_pt = readOGR("../data/field-spatial/point/cn_tr2_ee_pt.shp", "cn_tr2_ee_pt") 
```

# Extract NDVI with Pix4D
## 25cm - extract mean pixel values
### Mean
```{r}
extract_alnus_tr1_25cm = raster::extract(alnus_tr1_f08_ndvi, alnus_tr1_pt, buffer=.25, fun=mean, df=TRUE)
extract_bp_tr2_25cm = raster::extract(bp_tr2_f13_ndvi, bp_tr2_pt, buffer=.25, fun=mean,  df=TRUE)
extract_cn_tr1_25cm = raster::extract(cn_tr1_f20_ndvi, cn_tr1_pt, buffer=.25, fun=mean,  df=TRUE)
extract_cn_tr2_25cm = raster::extract(cn_tr2_f17_ndvi, cn_tr2_pt, buffer=.25, fun=mean,  df=TRUE)
```


```{r}
extract_alnus_tr1_25cm$plot_id = alnus_tr1_pt$ID
names(extract_alnus_tr1_25cm) = c('ID', 'mean_ndvi_25cm', 'plot_id')
extract_alnus_tr1_25cm
```

```{r}
extract_bp_tr2_25cm$plot_id = bp_tr2_pt$ID
names(extract_bp_tr2_25cm) = c('ID', 'mean_ndvi_25cm', 'plot_id')
extract_bp_tr2_25cm
```

```{r}
extract_cn_tr1_25cm$plot_id = cn_tr1_pt$ID
names(extract_cn_tr1_25cm) = c('ID', 'mean_ndvi_25cm', 'plot_id')
extract_cn_tr1_25cm
```

```{r}
extract_cn_tr2_25cm$plot_id = cn_tr2_pt$ID
names(extract_cn_tr2_25cm) = c('ID', 'mean_ndvi_25cm', 'plot_id')
extract_cn_tr2_25cm
```


```{r}
ndvi_25cm = bind_rows(extract_alnus_tr1_25cm,extract_bp_tr2_25cm, extract_cn_tr1_25cm, extract_cn_tr2_25cm) %>% dplyr::select(plot_id, mean_ndvi_25cm) %>% rename(ID = plot_id)
length(unique(ndvi_25cm$ID))
ndvi_25cm
```

### SD
```{r}
sd_alnus_tr1_25cm = raster::extract(alnus_tr1_f08_ndvi, alnus_tr1_pt, buffer=.25, fun=sd, df=TRUE)
sd_bp_tr2_25cm = raster::extract(bp_tr2_f13_ndvi, bp_tr2_pt, buffer=.25, fun=sd,  df=TRUE)
sd_cn_tr1_25cm = raster::extract(cn_tr1_f20_ndvi, cn_tr1_pt, buffer=.25, fun=sd,  df=TRUE)
sd_cn_tr2_25cm = raster::extract(cn_tr2_f17_ndvi, cn_tr2_pt, buffer=.25, fun=sd,  df=TRUE)
```


```{r}
sd_alnus_tr1_25cm$plot_id = alnus_tr1_pt$ID
names(sd_alnus_tr1_25cm) = c('ID', 'sd_ndvi_25cm', 'plot_id')
sd_alnus_tr1_25cm
```

```{r}
sd_bp_tr2_25cm$plot_id = bp_tr2_pt$ID
names(sd_bp_tr2_25cm) = c('ID', 'sd_ndvi_25cm', 'plot_id')
sd_bp_tr2_25cm
```

```{r}
sd_cn_tr1_25cm$plot_id = cn_tr1_pt$ID
names(sd_cn_tr1_25cm) = c('ID', 'sd_ndvi_25cm', 'plot_id')
sd_cn_tr1_25cm
```

```{r}
sd_cn_tr2_25cm$plot_id = cn_tr2_pt$ID
names(sd_cn_tr2_25cm) = c('ID', 'sd_ndvi_25cm', 'plot_id')
sd_cn_tr2_25cm
```


```{r}
ndvi_25cm_sd = bind_rows(sd_alnus_tr1_25cm,sd_bp_tr2_25cm, sd_cn_tr1_25cm, sd_cn_tr2_25cm) %>% dplyr::select(plot_id, sd_ndvi_25cm) %>% dplyr::rename(ID = plot_id)
ndvi_25cm_sd
```



## 50cm - extract mean pixel values
### Mean
```{r}
extract_alnus_tr1_50cm = raster::extract(alnus_tr1_f08_ndvi, alnus_tr1_pt, buffer=.50, fun=mean, df=TRUE)
extract_bp_tr2_50cm = raster::extract(bp_tr2_f13_ndvi, bp_tr2_pt, buffer=.50, fun=mean,  df=TRUE)
extract_cn_tr1_50cm = raster::extract(cn_tr1_f20_ndvi, cn_tr1_pt, buffer=.50, fun=mean,  df=TRUE)
extract_cn_tr2_50cm = raster::extract(cn_tr2_f17_ndvi, cn_tr2_pt, buffer=.50, fun=mean,  df=TRUE)
```


```{r}
extract_alnus_tr1_50cm$plot_id = alnus_tr1_pt$ID
names(extract_alnus_tr1_50cm) = c('ID', 'mean_ndvi_50cm', 'plot_id')
extract_alnus_tr1_50cm
```


```{r}
extract_bp_tr2_50cm$plot_id = bp_tr2_pt$ID
names(extract_bp_tr2_50cm) = c('ID', 'mean_ndvi_50cm', 'plot_id')
extract_bp_tr2_50cm
```

```{r}
extract_cn_tr1_50cm$plot_id = cn_tr1_pt$ID
names(extract_cn_tr1_50cm) = c('ID', 'mean_ndvi_50cm', 'plot_id')
extract_cn_tr1_50cm
```

```{r}
extract_cn_tr2_50cm$plot_id = cn_tr2_pt$ID
names(extract_cn_tr2_50cm) = c('ID', 'mean_ndvi_50cm', 'plot_id')
extract_cn_tr2_50cm
```


```{r}
ndvi_50cm = bind_rows(extract_alnus_tr1_50cm, extract_bp_tr2_50cm, extract_cn_tr1_50cm, extract_cn_tr2_50cm) %>% dplyr::select(plot_id, mean_ndvi_50cm) %>% rename(ID = plot_id)
length(unique(ndvi_50cm$ID))
ndvi_50cm
```

### SD
```{r}
sd_alnus_tr1_50cm = raster::extract(alnus_tr1_f08_ndvi, alnus_tr1_pt, buffer=.50, fun=sd, df=TRUE)
sd_bp_tr2_50cm = raster::extract(bp_tr2_f13_ndvi, bp_tr2_pt, buffer=.50, fun=sd,  df=TRUE)
sd_cn_tr1_50cm = raster::extract(cn_tr1_f20_ndvi, cn_tr1_pt, buffer=.50, fun=sd,  df=TRUE)
sd_cn_tr2_50cm = raster::extract(cn_tr2_f17_ndvi, cn_tr2_pt, buffer=.50, fun=sd,  df=TRUE)
```


```{r}
sd_alnus_tr1_50cm$plot_id = alnus_tr1_pt$ID
names(sd_alnus_tr1_50cm) = c('ID', 'sd_ndvi_50cm', 'plot_id')
sd_alnus_tr1_50cm
```


```{r}
sd_bp_tr2_50cm$plot_id = bp_tr2_pt$ID
names(sd_bp_tr2_50cm) = c('ID', 'sd_ndvi_50cm', 'plot_id')
sd_bp_tr2_50cm
```

```{r}
sd_cn_tr1_50cm$plot_id = cn_tr1_pt$ID
names(sd_cn_tr1_50cm) = c('ID', 'sd_ndvi_50cm', 'plot_id')
sd_cn_tr1_50cm
```

```{r}
sd_cn_tr2_50cm$plot_id = cn_tr2_pt$ID
names(sd_cn_tr2_50cm) = c('ID', 'sd_ndvi_50cm', 'plot_id')
sd_cn_tr2_50cm
```


```{r}
ndvi_50cm_sd = bind_rows(sd_alnus_tr1_50cm, sd_bp_tr2_50cm, sd_cn_tr1_50cm, sd_cn_tr2_50cm) %>% dplyr::select(plot_id, sd_ndvi_50cm) %>% dplyr::rename(ID = plot_id)
ndvi_50cm_sd
```


## 1m - extract mean pixel values
### Mean
```{r}
extract_alnus_tr1_1m = raster::extract(alnus_tr1_f08_ndvi, alnus_tr1_pt, buffer=1, fun=mean, df=TRUE)
extract_bp_tr2_1m = raster::extract(bp_tr2_f13_ndvi, bp_tr2_pt, buffer=1, fun=mean,  df=TRUE)
extract_cn_tr1_1m = raster::extract(cn_tr1_f20_ndvi, cn_tr1_pt, buffer=1, fun=mean,  df=TRUE)
extract_cn_tr2_1m = raster::extract(cn_tr2_f17_ndvi, cn_tr2_pt, buffer=1, fun=mean,  df=TRUE)
```


```{r}
extract_alnus_tr1_1m$plot_id = alnus_tr1_pt$ID
names(extract_alnus_tr1_1m) = c('ID', 'mean_ndvi_1m', 'plot_id')
extract_alnus_tr1_1m
```

```{r}
extract_bp_tr2_1m$plot_id = bp_tr2_pt$ID
names(extract_bp_tr2_1m) = c('ID', 'mean_ndvi_1m', 'plot_id')
extract_bp_tr2_1m
```

```{r}
extract_cn_tr1_1m$plot_id = cn_tr1_pt$ID
names(extract_cn_tr1_1m) = c('ID', 'mean_ndvi_1m', 'plot_id')
extract_cn_tr1_1m
```

```{r}
extract_cn_tr2_1m$plot_id = cn_tr2_pt$ID
names(extract_cn_tr2_1m) = c('ID', 'mean_ndvi_1m', 'plot_id')
extract_cn_tr2_1m
```


```{r}
ndvi_1m = bind_rows(extract_alnus_tr1_1m, extract_bp_tr2_1m, extract_cn_tr1_1m, extract_cn_tr2_1m) %>% dplyr::select(plot_id, mean_ndvi_1m) %>% rename(ID = plot_id)
length(unique(ndvi_1m$ID))
ndvi_1m
```

### SD
```{r}
sd_alnus_tr1_1m = raster::extract(alnus_tr1_f08_ndvi, alnus_tr1_pt, buffer=1, fun=sd, df=TRUE)
sd_bp_tr2_1m = raster::extract(bp_tr2_f13_ndvi, bp_tr2_pt, buffer=1, fun=sd,  df=TRUE)
sd_cn_tr1_1m = raster::extract(cn_tr1_f20_ndvi, cn_tr1_pt, buffer=1, fun=sd,  df=TRUE)
sd_cn_tr2_1m = raster::extract(cn_tr2_f17_ndvi, cn_tr2_pt, buffer=1, fun=sd,  df=TRUE)
```


```{r}
sd_alnus_tr1_1m$plot_id = alnus_tr1_pt$ID
names(sd_alnus_tr1_1m) = c('ID', 'sd_ndvi_1m', 'plot_id')
sd_alnus_tr1_1m
```

```{r}
sd_bp_tr2_1m$plot_id = bp_tr2_pt$ID
names(sd_bp_tr2_1m) = c('ID', 'sd_ndvi_1m', 'plot_id')
sd_bp_tr2_1m
```

```{r}
sd_cn_tr1_1m$plot_id = cn_tr1_pt$ID
names(sd_cn_tr1_1m) = c('ID', 'sd_ndvi_1m', 'plot_id')
sd_cn_tr1_1m
```

```{r}
sd_cn_tr2_1m$plot_id = cn_tr2_pt$ID
names(sd_cn_tr2_1m) = c('ID', 'sd_ndvi_1m', 'plot_id')
sd_cn_tr2_1m
```


```{r}
ndvi_1m_sd = bind_rows(sd_alnus_tr1_1m, sd_bp_tr2_1m, sd_cn_tr1_1m, sd_cn_tr2_1m) %>% dplyr::select(plot_id,  sd_ndvi_1m) %>% dplyr::rename(ID = plot_id)
ndvi_1m_sd
```



## 3m - extract mean pixel values
### Mean
```{r}
extract_alnus_tr1_3m = raster::extract(alnus_tr1_f08_ndvi, alnus_tr1_pt, buffer=3, fun=mean, df=TRUE)
extract_bp_tr2_3m = raster::extract(bp_tr2_f13_ndvi, bp_tr2_pt, buffer=3, fun=mean,  df=TRUE)
extract_cn_tr1_3m = raster::extract(cn_tr1_f20_ndvi, cn_tr1_pt, buffer=3, fun=mean,  df=TRUE)
extract_cn_tr2_3m = raster::extract(cn_tr2_f17_ndvi, cn_tr2_pt, buffer=3, fun=mean,  df=TRUE)
```


```{r}
extract_alnus_tr1_3m$plot_id = alnus_tr1_pt$ID
names(extract_alnus_tr1_3m) = c('ID', 'mean_ndvi_3m', 'plot_id')
extract_alnus_tr1_3m
```

```{r}
extract_bp_tr2_3m$plot_id = bp_tr2_pt$ID
names(extract_bp_tr2_3m) = c('ID', 'mean_ndvi_3m', 'plot_id')
extract_bp_tr2_3m
```

```{r}
extract_cn_tr1_3m$plot_id = cn_tr1_pt$ID
names(extract_cn_tr1_3m) = c('ID', 'mean_ndvi_3m', 'plot_id')
extract_cn_tr1_3m
```

```{r}
extract_cn_tr2_3m$plot_id = cn_tr2_pt$ID
names(extract_cn_tr2_3m) = c('ID', 'mean_ndvi_3m', 'plot_id')
extract_cn_tr2_3m
```


```{r}
ndvi_3m = bind_rows(extract_alnus_tr1_3m, extract_bp_tr2_3m, extract_cn_tr1_3m, extract_cn_tr2_3m) %>% dplyr::select(plot_id, mean_ndvi_3m) %>% rename(ID = plot_id)
length(unique(ndvi_3m$ID))
ndvi_3m
```

### SD
```{r}
sd_alnus_tr1_3m = raster::extract(alnus_tr1_f08_ndvi, alnus_tr1_pt, buffer=3, fun=sd, df=TRUE)
sd_bp_tr2_3m = raster::extract(bp_tr2_f13_ndvi, bp_tr2_pt, buffer=3, fun=sd,  df=TRUE)
sd_cn_tr1_3m = raster::extract(cn_tr1_f20_ndvi, cn_tr1_pt, buffer=3, fun=sd,  df=TRUE)
sd_cn_tr2_3m = raster::extract(cn_tr2_f17_ndvi, cn_tr2_pt, buffer=3, fun=sd,  df=TRUE)
```


```{r}
sd_alnus_tr1_3m$plot_id = alnus_tr1_pt$ID
names(sd_alnus_tr1_3m) = c('ID', 'sd_ndvi_3m', 'plot_id')
sd_alnus_tr1_3m
```

```{r}
sd_bp_tr2_3m$plot_id = bp_tr2_pt$ID
names(sd_bp_tr2_3m) = c('ID', 'sd_ndvi_3m', 'plot_id')
sd_bp_tr2_3m
```

```{r}
sd_cn_tr1_3m$plot_id = cn_tr1_pt$ID
names(sd_cn_tr1_3m) = c('ID', 'sd_ndvi_3m', 'plot_id')
sd_cn_tr1_3m
```

```{r}
sd_cn_tr2_3m$plot_id = cn_tr2_pt$ID
names(sd_cn_tr2_3m) = c('ID', 'sd_ndvi_3m', 'plot_id')
sd_cn_tr2_3m
```


```{r}
ndvi_3m_sd = bind_rows(sd_alnus_tr1_3m, sd_bp_tr2_3m, sd_cn_tr1_3m, sd_cn_tr2_3m) %>% dplyr::select(plot_id, sd_ndvi_3m) %>% dplyr::rename(ID = plot_id)
ndvi_3m_sd
```



## 5m - extract mean pixel values
### Mean
```{r}
extract_alnus_tr1_5m = raster::extract(alnus_tr1_f08_ndvi, alnus_tr1_pt, buffer=5, fun=mean, df=TRUE)
extract_bp_tr2_5m = raster::extract(bp_tr2_f13_ndvi, bp_tr2_pt, buffer=5, fun=mean,  df=TRUE)
extract_cn_tr1_5m = raster::extract(cn_tr1_f20_ndvi, cn_tr1_pt, buffer=5, fun=mean,  df=TRUE)
extract_cn_tr2_5m = raster::extract(cn_tr2_f17_ndvi, cn_tr2_pt, buffer=5, fun=mean,  df=TRUE)
```


```{r}
extract_alnus_tr1_5m$plot_id = alnus_tr1_pt$ID
names(extract_alnus_tr1_5m) = c('ID', 'mean_ndvi_5m', 'plot_id')
extract_alnus_tr1_5m
```

```{r}
extract_bp_tr2_5m$plot_id = bp_tr2_pt$ID
names(extract_bp_tr2_5m) = c('ID', 'mean_ndvi_5m', 'plot_id')
extract_bp_tr2_5m
```

```{r}
extract_cn_tr1_5m$plot_id = cn_tr1_pt$ID
names(extract_cn_tr1_5m) = c('ID', 'mean_ndvi_5m', 'plot_id')
extract_cn_tr1_5m
```

```{r}
extract_cn_tr2_5m$plot_id = cn_tr2_pt$ID
names(extract_cn_tr2_5m) = c('ID', 'mean_ndvi_5m', 'plot_id')
extract_cn_tr2_5m
```


```{r}
ndvi_5m = bind_rows(extract_alnus_tr1_5m, extract_bp_tr2_5m,extract_cn_tr1_5m, extract_cn_tr2_5m) %>% dplyr::select(plot_id, mean_ndvi_5m) %>% rename(ID = plot_id)
length(unique(ndvi_5m$ID))
ndvi_5m
```

### SD
```{r}
sd_alnus_tr1_5m = raster::extract(alnus_tr1_f08_ndvi, alnus_tr1_pt, buffer=5, fun=sd, df=TRUE)
sd_bp_tr2_5m = raster::extract(bp_tr2_f13_ndvi, bp_tr2_pt, buffer=5, fun=sd,  df=TRUE)
sd_cn_tr1_5m = raster::extract(cn_tr1_f20_ndvi, cn_tr1_pt, buffer=5, fun=sd,  df=TRUE)
sd_cn_tr2_5m = raster::extract(cn_tr2_f17_ndvi, cn_tr2_pt, buffer=5, fun=sd,  df=TRUE)
```


```{r}
sd_alnus_tr1_5m$plot_id = alnus_tr1_pt$ID
names(sd_alnus_tr1_5m) = c('ID', 'sd_ndvi_5m', 'plot_id')
sd_alnus_tr1_5m
```

```{r}
sd_bp_tr2_5m$plot_id = bp_tr2_pt$ID
names(sd_bp_tr2_5m) = c('ID', 'sd_ndvi_5m', 'plot_id')
sd_bp_tr2_5m
```

```{r}
sd_cn_tr1_5m$plot_id = cn_tr1_pt$ID
names(sd_cn_tr1_5m) = c('ID', 'sd_ndvi_5m', 'plot_id')
sd_cn_tr1_5m
```

```{r}
sd_cn_tr2_5m$plot_id = cn_tr2_pt$ID
names(sd_cn_tr2_5m) = c('ID', 'sd_ndvi_5m', 'plot_id')
sd_cn_tr2_5m
```


```{r}
ndvi_5m_sd = bind_rows(sd_alnus_tr1_5m, sd_bp_tr2_5m,sd_cn_tr1_5m, sd_cn_tr2_5m) %>% dplyr::select(plot_id, sd_ndvi_5m) %>% dplyr::rename(ID = plot_id)

ndvi_5m_sd
```


## 10m - extract mean pixel values
###Mean
```{r}
extract_alnus_tr1_10m = raster::extract(alnus_tr1_f08_ndvi, alnus_tr1_pt, buffer=10, fun=mean, df=TRUE)
extract_bp_tr2_10m = raster::extract(bp_tr2_f13_ndvi, bp_tr2_pt, buffer=10, fun=mean,  df=TRUE)
extract_cn_tr1_10m = raster::extract(cn_tr1_f20_ndvi, cn_tr1_pt, buffer=10, fun=mean,  df=TRUE)
extract_cn_tr2_10m = raster::extract(cn_tr2_f17_ndvi, cn_tr2_pt, buffer=10, fun=mean,  df=TRUE)
```


```{r}
extract_alnus_tr1_10m$plot_id = alnus_tr1_pt$ID
names(extract_alnus_tr1_10m) = c('ID', 'mean_ndvi_10m', 'plot_id')
extract_alnus_tr1_10m
```

```{r}
extract_bp_tr2_10m$plot_id = bp_tr2_pt$ID
names(extract_bp_tr2_10m) = c('ID', 'mean_ndvi_10m', 'plot_id')
extract_bp_tr2_10m
```

```{r}
extract_cn_tr1_10m$plot_id = cn_tr1_pt$ID
names(extract_cn_tr1_10m) = c('ID', 'mean_ndvi_10m', 'plot_id')
extract_cn_tr1_10m
```

```{r}
extract_cn_tr2_10m$plot_id = cn_tr2_pt$ID
names(extract_cn_tr2_10m) = c('ID', 'mean_ndvi_10m', 'plot_id')
extract_cn_tr2_10m
```

```{r}
ndvi_10m = bind_rows(extract_alnus_tr1_10m, extract_bp_tr2_10m, extract_cn_tr1_10m, extract_cn_tr2_10m) %>% dplyr::select(plot_id, mean_ndvi_10m) %>% rename(ID = plot_id)
length(unique(ndvi_10m$ID))
ndvi_10m
```

### SD
```{r}
sd_alnus_tr1_10m = raster::extract(alnus_tr1_f08_ndvi, alnus_tr1_pt, buffer=10, fun=sd, df=TRUE)
sd_bp_tr2_10m = raster::extract(bp_tr2_f13_ndvi, bp_tr2_pt, buffer=10, fun=sd,  df=TRUE)
sd_cn_tr1_10m = raster::extract(cn_tr1_f20_ndvi, cn_tr1_pt, buffer=10, fun=sd,  df=TRUE)
sd_cn_tr2_10m = raster::extract(cn_tr2_f17_ndvi, cn_tr2_pt, buffer=10, fun=sd,  df=TRUE)
```


```{r}
sd_alnus_tr1_10m$plot_id = alnus_tr1_pt$ID
names(sd_alnus_tr1_10m) = c('ID', 'sd_ndvi_10m', 'plot_id')
sd_alnus_tr1_10m
```

```{r}
sd_bp_tr2_10m$plot_id = bp_tr2_pt$ID
names(sd_bp_tr2_10m) = c('ID', 'sd_ndvi_10m', 'plot_id')
sd_bp_tr2_10m
```

```{r}
sd_cn_tr1_10m$plot_id = cn_tr1_pt$ID
names(sd_cn_tr1_10m) = c('ID', 'sd_ndvi_10m', 'plot_id')
sd_cn_tr1_10m
```

```{r}
sd_cn_tr2_10m$plot_id = cn_tr2_pt$ID
names(sd_cn_tr2_10m) = c('ID', 'sd_ndvi_10m', 'plot_id')
sd_cn_tr2_10m
```

```{r}
ndvi_10m_sd = bind_rows(sd_alnus_tr1_10m, sd_bp_tr2_10m, sd_cn_tr1_10m, sd_cn_tr2_10m) %>% dplyr::select(plot_id, sd_ndvi_10m) %>% dplyr::rename(ID = plot_id)
ndvi_10m_sd
```

## Tree - extract mean pixel values

```{r}
extract_alnus_tr1_tree = raster::extract(alnus_tr1_f08_ndvi, alnus_tr1_pt, buffer=alnus_tr1_pt$radius, fun=mean, df=TRUE)
extract_bp_tr2_tree = raster::extract(bp_tr2_f13_ndvi, bp_tr2_pt, buffer=bp_tr2_pt$radius, fun=mean,  df=TRUE)
extract_cn_tr1_tree = raster::extract(cn_tr1_f20_ndvi, cn_tr1_pt, buffer=cn_tr1_pt$radius, fun=mean,  df=TRUE)
extract_cn_tr2_tree = raster::extract(cn_tr2_f17_ndvi, cn_tr2_pt, buffer=cn_tr2_pt$radius, fun=mean,  df=TRUE)
```


```{r}
extract_alnus_tr1_tree$plot_id = alnus_tr1_pt$ID
names(extract_alnus_tr1_tree) = c('ID', 'mean_ndvi_tree', 'plot_id')
extract_alnus_tr1_tree
```

```{r}
extract_bp_tr2_tree$plot_id = bp_tr2_pt$ID
names(extract_bp_tr2_tree) = c('ID', 'mean_ndvi_tree', 'plot_id')
extract_bp_tr2_tree
```

```{r}
extract_cn_tr1_tree$plot_id = cn_tr1_pt$ID
names(extract_cn_tr1_tree) = c('ID', 'mean_ndvi_tree', 'plot_id')
extract_cn_tr1_tree
```

```{r}
extract_cn_tr2_tree$plot_id = cn_tr2_pt$ID
names(extract_cn_tr2_tree) = c('ID', 'mean_ndvi_tree', 'plot_id')
extract_cn_tr2_tree
```


```{r}
ndvi_tree = bind_rows(extract_alnus_tr1_tree, extract_bp_tr2_tree, extract_cn_tr1_tree, extract_cn_tr2_tree) %>% dplyr::select(plot_id, mean_ndvi_tree) %>% rename(ID = plot_id)
length(unique(ndvi_tree$ID))
ndvi_tree
```

## shrub - extract mean pixel values
```{r}
extract_alnus_tr1_shrub = raster::extract(alnus_tr1_f08_ndvi, alnus_tr1_pt, buffer=alnus_tr1_pt$plt_lng, fun=mean, df=TRUE)
extract_bp_tr2_shrub = raster::extract(bp_tr2_f13_ndvi, bp_tr2_pt, buffer=cn_tr2_pt$plt_lng, fun=mean,  df=TRUE)
extract_cn_tr1_shrub = raster::extract(cn_tr1_f20_ndvi, cn_tr1_pt, buffer=cn_tr1_pt$plt_lng, fun=mean,  df=TRUE)
extract_cn_tr2_shrub = raster::extract(cn_tr2_f17_ndvi, cn_tr2_pt, buffer=cn_tr2_pt$plt_lng, fun=mean,  df=TRUE)
```


```{r}
extract_alnus_tr1_shrub$plot_id = alnus_tr1_pt$ID
names(extract_alnus_tr1_shrub) = c('ID', 'mean_ndvi_shrub', 'plot_id')
extract_alnus_tr1_shrub
```

```{r}
extract_bp_tr2_shrub$plot_id = bp_tr2_pt$ID
names(extract_bp_tr2_shrub) = c('ID', 'mean_ndvi_shrub', 'plot_id')
extract_bp_tr2_shrub
```

```{r}
extract_cn_tr1_shrub$plot_id = cn_tr1_pt$ID
names(extract_cn_tr1_shrub) = c('ID', 'mean_ndvi_shrub', 'plot_id')
extract_cn_tr1_shrub
```

```{r}
extract_cn_tr2_shrub$plot_id = cn_tr2_pt$ID
names(extract_cn_tr2_shrub) = c('ID', 'mean_ndvi_shrub', 'plot_id')
extract_cn_tr2_shrub
```


```{r}
ndvi_shrub = bind_rows(extract_alnus_tr1_shrub, extract_bp_tr2_shrub, extract_cn_tr1_shrub, extract_cn_tr2_shrub) %>% dplyr::select(plot_id, mean_ndvi_shrub) %>% rename(ID = plot_id)
length(unique(ndvi_shrub$ID))
ndvi_shrub
```


## Combine ndvi data

###Mean
```{r}
ndvi_merge1 = merge(ndvi_25cm, ndvi_50cm,  by= "ID")
ndvi_merge2 = merge(ndvi_1m, ndvi_3m,  by= "ID")
ndvi_merge3 = merge(ndvi_5m, ndvi_10m,  by= "ID")
ndvi_merge4 = merge(ndvi_tree, ndvi_shrub,  by="ID")
ndvi_merge5 = merge(ndvi_merge1, ndvi_merge2, by="ID")
ndvi_merge6 = merge(ndvi_merge3, ndvi_merge4, by="ID")
ndvi_all = merge(ndvi_merge5, ndvi_merge6,  by="ID")
ndvi_all
```


```{r}
write.csv(ndvi_all,"../outputs/extracted_data/2020-08-10_ndvi_all_pix4D.csv", row.names = FALSE)
```


### SD
```{r}
ndvi_merge1_sd = merge(ndvi_25cm_sd, ndvi_50cm_sd,  by= "ID")
ndvi_merge2_sd = merge(ndvi_1m_sd, ndvi_3m_sd,  by= "ID")
ndvi_merge3_sd = merge(ndvi_5m_sd, ndvi_10m_sd,  by= "ID")
ndvi_merge5_sd = merge(ndvi_merge1_sd, ndvi_merge2_sd, by="ID")
ndvi_all_sd = merge(ndvi_merge5_sd, ndvi_merge3_sd, by="ID")
ndvi_all_sd
```


```{r}
write.csv(ndvi_all_sd,"../outputs/extracted_data/2020-08-18_ndvi_all_sd.csv", row.names = FALSE)
```



# Extract all values within buffer

## 10m - extract all pixel values

```{r}
extract_alnus_tr1_10m_all = raster::extract(alnus_tr1_f08_ndvi, alnus_tr1_pt, buffer=10, df=TRUE)
```

```{r}
extract_bp_tr2_10m_all = raster::extract(bp_tr2_f13_ndvi, bp_tr2_pt, buffer=10, df=TRUE)

extract_cn_tr1_10m_all = raster::extract(cn_tr1_f20_ndvi, cn_tr1_pt, buffer=10, df=TRUE)

extract_cn_tr2_10m_all = raster::extract(cn_tr2_f17_ndvi, cn_tr2_pt, buffer=10, df=TRUE)
```

```{r}
head(extract_alnus_tr1_10m_all)
tail(extract_alnus_tr1_10m_all)
```
### alnus add plot id
```{r}
alnus_tr1_pt@data
```
```{r}
df_extract_alnus_tr1_10m_all = data.frame(extract_alnus_tr1_10m_all)
```


```{r}
alnus_ndvi = df_extract_alnus_tr1_10m_all %>%
  mutate(plot_id = ifelse(ID == "1", "Alnus_1984_1_-25",
                    ifelse(ID == "2", "Alnus_1984_1_-50",
                    ifelse(ID == "3", "Alnus_1984_1_100",
                    ifelse(ID == "4", "Alnus_1984_1_150", 
                    ifelse(ID == "5", "Alnus_1984_1_200",
                    ifelse(ID == "6", "Alnus_1984_1_25",
                    ifelse(ID == "7", "Alnus_1984_1_50", "Alnus_1984_1_75")))))))) %>%
  rename(ndvi = RU_ALN_TR1_FL008M_ndvi)
```


### bp tr2 add plot id
```{r}
bp_tr2_pt@data
```
```{r}
df_extract_bp_tr2_10m_all = data.frame(extract_bp_tr2_10m_all)
head(df_extract_bp_tr2_10m_all)
```


```{r}
bptr2_ndvi = df_extract_bp_tr2_10m_all %>%
  mutate(plot_id = ifelse(ID == "1", "BP_1983_2_-25a",
                    ifelse(ID == "2", "BP_1983_2_-25b",
                    ifelse(ID == "3", "BP_1983_2_100",
                    ifelse(ID == "4", "BP_1983_2_150", 
                    ifelse(ID == "5", "BP_1983_2_200",
                    ifelse(ID == "6", "BP_1983_2_25",
                    ifelse(ID == "7", "BP_1983_2_50", "BP_1983_2_75")))))))) %>%
  rename(ndvi = RU_BP_TR2_FL013M_ndvi)

head(bptr2_ndvi)
```



### CN tr1 add plot id
```{r}
cn_tr1_pt@data
```
```{r}
df_extract_cn_tr1_10m_all = data.frame(extract_cn_tr1_10m_all)
head(df_extract_cn_tr1_10m_all)
```


```{r}
cntr1_ndvi = df_extract_cn_tr1_10m_all %>%
  mutate(plot_id = ifelse(ID == "1", "CN_2001_1_-25",
                    ifelse(ID == "2", "CN_2001_1_-50",
                    ifelse(ID == "3", "CN_2001_1_100",
                    ifelse(ID == "4", "CN_2001_1_150", 
                    ifelse(ID == "5", "CN_2001_1_200",
                    ifelse(ID == "6", "CN_2001_1_25",
                    ifelse(ID == "7", "CN_2001_1_50", "CN_2001_1_75")))))))) %>%
  rename(ndvi = RU_CYN_TR1_FL020M_ndvi)

head(cntr1_ndvi)
```



### CN tr2 add plot id
```{r}
cn_tr2_pt@data
```
```{r}
df_extract_cn_tr2_10m_all = data.frame(extract_cn_tr2_10m_all)
head(df_extract_cn_tr2_10m_all)
```


```{r}
cntr2_ndvi = df_extract_cn_tr2_10m_all %>%
  mutate(plot_id = ifelse(ID == "1", "CN_2001_2_-35",
                    ifelse(ID == "2", "CN_2001_2_-60",
                    ifelse(ID == "3", "CN_2001_2_100",
                    ifelse(ID == "4", "CN_2001_2_150", 
                    ifelse(ID == "5", "CN_2001_2_200",
                    ifelse(ID == "6", "CN_2001_2_25",
                    ifelse(ID == "7", "CN_2001_2_50", "CN_2001_2_75")))))))) %>%
  rename(ndvi = RU_CYN_TR2_FL017M_ndvi)

head(cntr2_ndvi)
```

## Combine in to single data frame

```{r}
ndvi_10m_all =  bind_rows(alnus_ndvi, bptr2_ndvi, cntr1_ndvi, cntr2_ndvi) 
summary(ndvi_10m_all)
```

```{r}
write.csv(ndvi_10m_all,"../outputs/extracted_data/2020-08-13_ndvi_all_pixels.csv", row.names = FALSE)
```
