==> scripts/2.4_UAV-Field-Raw-Data-Visualization.Rmd <==
---
title: "UAV-Field-Raw-Data-Visualization"
author: "Anna Talucci"
date: "5/13/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```
# Overview


# Pakcages
```{r}
library(tidyr)
library(dplyr)
library(ggplot2)
library(stringr)
library(lme4)
library(cowplot)
library(lme4)
library(lsmeans)
library(DHARMa)
```


# Field Data
```{r}
field_data = read.csv("../data/field_subset/field_data_all.csv")
```

```{r}
head(field_data)
```

## field Plot sizes

```{r}
field_data %>% group_by(site, transect) %>% 
  summarise(min(shrub_plot_area))
```

```{r}
field_data %>% group_by(site, transect) %>% 
  summarise(max(shrub_plot_area))
```

```{r}
field_data %>% group_by(site, transect) %>% 
  summarise(min(tree_radius))
```

```{r}
field_data %>% group_by(site, transect) %>% 
  summarise(max(tree_radius))
```

# UAV Data

```{r}
gcc_all = read.csv("../data/extracted_data/gcc_all.csv", strip.white = TRUE)
ndvi_all = read.csv("../data/extracted_data/ndvi_all.csv", strip.white = TRUE)
```

```{r}
head(gcc_all)
head(ndvi_all)
```

# Combine Field and UAV data

```{r}
gcc_ndvi = merge(gcc_all, ndvi_all, by="ID")
gcc_ndvi
field_gcc = merge(field_data, gcc_all, by="ID") 
field_gcc
field_gcc_ndvi = merge(field_data, gcc_ndvi,  "ID")
field_gcc_ndvi
```

# burn_year as factor

```{r}
field_gcc$burn_year = as.factor(field_gcc$burn_year)
field_gcc_ndvi$burn_year = as.factor(field_gcc_ndvi$burn_year)

str(field_gcc)
str(field_gcc_ndvi)
```



# Define graph Theme and color Palette 

```{r}
comp_theme = theme_bw() + theme(legend.position = "none") +  
  theme(panel.grid.major = element_blank()) +
    theme(axis.title.y = element_text(size = 11, hjust = 0.5, vjust = 1.1),
        axis.text.x = element_text(size = 10, color = "black"),
        axis.text.y = element_text(size = 10, color = "black"),
        axis.line = element_line(colour = "black"))
```
 
```{r}

reg_theme = theme_bw() + theme(legend.position = "none") +
  theme(plot.margin = unit(c(t = 0.3, r = 0.3, b = 0.3, l = 0.1), "cm")) +
  theme(axis.title.x = element_text(size = 11, hjust = 0.5, vjust = -0.1),
        axis.title.y = element_text(size = 11, hjust = 0.5, vjust = 1.1),
        axis.text.x = element_text(size = 10, color = "black"),
        axis.text.y = element_text(size = 10, color = "black"),
        panel.background = element_blank(), 
        axis.line = element_line(colour = "black"))

```

```{r}
cbbPalette <- c("#000000", "#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#D55E00", "#CC79A7")
```

# Data visualization


```{r}
ggplot(data = field_gcc, aes(x = plot_dis , y = mean_gcc_25cm, color = factor(transect_ID))) +
         geom_point(size = 3) +
         scale_color_manual(values=cbbPalette) +
          ylab("mean gcc (25cm)") +
          xlab("distance from edge")  +
          theme_bw() 
```

```{r}
ggplot(data = field_gcc, aes(x = transect_ID , y = shrub_total_ba, color = factor(transect_ID))) +
  geom_boxplot() +       
  geom_point() +
         scale_color_manual(values=cbbPalette) +
          ylab("shrub basal area") +
          xlab("Transect ID")  +
           theme_bw() + theme(legend.position = "bottom") +
  theme(plot.margin = unit(c(t = 0.3, r = 0.1, b = 0.3, l = 0.1), "cm")) +
  theme(axis.title.x = element_text(size = 11, hjust = 0.5, vjust = -0.1),
        axis.title.y = element_text(size = 11, hjust = 0.5, vjust = 1.1),
        axis.text.x = element_blank(),
        axis.text.y = element_text(size = 10, color = "black"),
        panel.background = element_blank(), 
        axis.line = element_line(colour = "black")) + guides(color=guide_legend(override.aes=list(fill=NA), title = "Transect ID"))
```

```{r}
ggplot(data = field_gcc, aes(x = transect_ID , y = tree_total_ba, color = factor(transect_ID))) +
  geom_boxplot() +       
  geom_point() +
         scale_color_manual(values=cbbPalette) +
          ylab("Tree basal area") +
          xlab("Transect ID")  +
           theme_bw() + theme(legend.position = "bottom") +
  theme(plot.margin = unit(c(t = 0.3, r = 0.1, b = 0.3, l = 0.1), "cm")) +
  theme(axis.title.x = element_text(size = 11, hjust = 0.5, vjust = -0.1),
        axis.title.y = element_text(size = 11, hjust = 0.5, vjust = 1.1),
        axis.text.x = element_blank(),
        axis.text.y = element_text(size = 10, color = "black"),
        panel.background = element_blank(), 
        axis.line = element_line(colour = "black")) + guides(color=guide_legend(override.aes=list(fill=NA), title = "Transect ID"))
```

```{r}
ggplot(data = field_gcc, aes(x = transect_ID , y = prop_live, color = factor(transect_ID))) +
  geom_boxplot() +       
  geom_point() +
         scale_color_manual(values=cbbPalette) +
          ylab("Proportion live trees") +
          xlab("Transect ID")  +
           theme_bw() + theme(legend.position = "bottom") +
  theme(plot.margin = unit(c(t = 0.3, r = 0.1, b = 0.3, l = 0.1), "cm")) +
  theme(axis.title.x = element_text(size = 11, hjust = 0.5, vjust = -0.1),
        axis.title.y = element_text(size = 11, hjust = 0.5, vjust = 1.1),
        axis.text.x = element_blank(),
        axis.text.y = element_text(size = 10, color = "black"),
        panel.background = element_blank(), 
        axis.line = element_line(colour = "black")) + guides(color=guide_legend(override.aes=list(fill=NA), title = "Transect ID"))
```


Boxplot with points
```{r}
ggplot(data = field_gcc, aes(x = transect_ID , y = mean_gcc_25cm, color = factor(transect_ID))) +
  geom_boxplot() +       
  geom_point() +
         scale_color_manual(values=cbbPalette) +
          ylab("mean gcc (25cm)") +
          xlab("Transect ID")  +
           theme_bw() + theme(legend.position = "bottom") +
  theme(plot.margin = unit(c(t = 0.3, r = 0.1, b = 0.3, l = 0.1), "cm")) +
  theme(axis.title.x = element_text(size = 11, hjust = 0.5, vjust = -0.1),
        axis.title.y = element_text(size = 11, hjust = 0.5, vjust = 1.1),
        axis.text.x = element_blank(),
        axis.text.y = element_text(size = 10, color = "black"),
        panel.background = element_blank(), 
        axis.line = element_line(colour = "black")) + guides(color=guide_legend(override.aes=list(fill=NA), title = "Transect ID"))
```
```{r}
ggplot(data = field_gcc_ndvi, aes(x = mean_ndvi_25cm , y = mean_ndvi_25cm, color = factor(transect_ID))) +
         geom_point(size = 3) +
         scale_color_manual(values=cbbPalette) +
          ylab("mean ndvi (25cm)") +
          xlab("")  +
          theme_bw() 
```

```{r}
ggplot(data = field_gcc_ndvi, aes(x = mean_gcc_25cm , y = mean_ndvi_25cm, color = factor(transect_ID))) +
         geom_point(size = 3) +
         scale_color_manual(values=cbbPalette) +
          ylab("mean ndvi (25cm)") +
          xlab("mean gcc (25cm)")  +
          theme_bw() 
```

```{r}
ggplot(data = field_gcc, aes(x = plot_dis , y = mean_gcc_1m, color = factor(burn_year))) +
         geom_point(size = 3) +
         scale_color_manual(values=c("#000000", "#E69F00", "#56B4E9", "#009E73")) +
          ylab("mean gcc (1m)") +
          xlab("distance from edge")  +
          theme_bw() 
```

```{r}
ggplot(data = field_gcc, aes(x = plot_dis , y = mean_gcc_3m, color = factor(burn_year))) +
         geom_point(size = 3) +
         scale_color_manual(values=c("#000000", "#E69F00", "#56B4E9", "#009E73")) +
          ylab("mean gcc (3m)") +
          xlab("distance from edge")  +
          theme_bw() 
```

```{r}
ggplot(data = field_gcc, aes(x = prop_live , y = mean_gcc_tree, color = factor(burn_year))) +
         geom_point(size = 3) +
         scale_color_manual(values=c("#000000", "#E69F00", "#56B4E9", "#009E73")) +
          ylab("mean gcc") +
          xlab("proportion live trees")  +
          theme_bw() 
```


```{r}
ggplot(data = field_gcc, aes(x = shrub_ct , y = mean_gcc_shrub, color = factor(burn_year))) +
         geom_point(size = 3) +
         scale_color_manual(values=c("#000000", "#E69F00", "#56B4E9", "#009E73")) +
          ylab("mean gcc") +
          xlab("shrub abundance")  +
          theme_bw() 
```
