==> scripts/3.3_UAV-Biomass-Analysis.Rmd <==
---
title: "UAV-Biomass-Analysis"
author: "Anna Talucci"
date: "5/13/2020"
output: html_document
---



```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Objective
Evaluate how well vegetation indices predict aboveground carbon biomass.


# Pakcages
```{r}
library(tidyverse)
library(broom)
library(ggplot2)
library(stringr)
library(lme4)
library(cowplot)
library(lme4)
library(lsmeans)
library(magrittr)
library(rcompanion)
library(qpcR)

```

# Functions

```{r}
numextract <- function(string){ 
  str_extract(string, "\\-*\\d+\\.*\\d*")
} 

```

# Data
## Field Data
```{r}
field_data = read.csv("../data/field_subset/field_data_all.csv")
```

```{r}
head(field_data)
```

### field Plot sizes

```{r}
field_data %>% group_by(site, transect) %>% 
  summarise(min(shrub_plot_area))
```

```{r}
field_data %>% group_by(site, transect) %>% 
  summarise(max(shrub_plot_area))
```

```{r}
field_data %>% group_by(site, transect) %>% 
  summarise(min(tree_radius))
```

```{r}
field_data %>% group_by(site, transect) %>% 
  summarise(max(tree_radius))
```

## UAV Data

```{r}
gcc_all = read.csv("../data/extracted_data/gcc_all.csv", strip.white = TRUE)
ndvi_all = read.csv("../data/extracted_data/2020-08-10_ndvi_all_pix4D.csv", strip.white = TRUE)
```

```{r}
head(gcc_all)
head(ndvi_all)
```

## Pull out GCC and NDVI by buffer size

```{r}
gcc_long = gcc_all %>% 
  dplyr::select(ID:mean_gcc_10m) %>%
  pivot_longer(-ID, names_to = "buffer", values_to = "gcc") %>%
  mutate(buf_res = numextract(buffer))

gcc_long
```

```{r}
ndvi_long = ndvi_all %>% 
  dplyr::select(ID:mean_ndvi_10m) %>%
  pivot_longer(-ID, names_to = "buffer", values_to = "ndvi") %>%
  mutate(buf_res = numextract(buffer))

ndvi_long
```

## Combine Field and UAV data

```{r}
gcc_ndvi = merge(gcc_all, ndvi_all, by="ID")
gcc_ndvi
field_gcc = merge(field_data, gcc_all, by="ID") 
field_gcc
field_gcc_ndvi = merge(field_data, gcc_ndvi,  "ID")
field_gcc_ndvi
```

## Make DF long format
Convert data frame of Field data, GCC nad NDVI into a long format.
```{r}
field_gcc_long = merge(field_data, gcc_long, by="ID")

field_gcc_long1 = field_gcc_long %>% 
  dplyr::select(ID:burn_year, live_woody_biomass_C_g_m, buffer:buf_res) %>%
  mutate(res_meters = ifelse(buf_res %in% 25, "25cm",
                                    ifelse(buf_res %in% 50, "50cm",
                                    ifelse(buf_res %in% 1, "1m",
                                    ifelse(buf_res %in% 3, "3m",
                                    ifelse(buf_res %in% 5, "5m", "10m")))))) %>%
  mutate(res_meters = factor(res_meters, levels = c("25cm", "50cm", "1m", "3m", "5m", "10m"))) %>%
  mutate(biomass_log = log(live_woody_biomass_C_g_m))

head(field_gcc_long1)
```

```{r}
field_ndvi_long = merge(field_data, ndvi_long, by="ID")

field_ndvi_long1 = field_ndvi_long %>% 
  dplyr::select(ID:burn_year, live_woody_biomass_C_g_m, buffer:buf_res) %>%
  mutate(res_meters = ifelse(buf_res %in% 25, "25cm",
                                    ifelse(buf_res %in% 50, "50cm",
                                    ifelse(buf_res %in% 1, "1m",
                                    ifelse(buf_res %in% 3, "3m",
                                    ifelse(buf_res %in% 5, "5m", "10m")))))) %>%
  mutate(res_meters = factor(res_meters, levels = c("25cm", "50cm", "1m", "3m", "5m", "10m"))) %>%
  mutate(biomass_log = log(live_woody_biomass_C_g_m))

head(field_ndvi_long1)
```

## Variables as factor

```{r}
field_gcc$burn_year = as.factor(field_gcc$burn_year)
field_gcc_ndvi$burn_year = as.factor(field_gcc_ndvi$burn_year)


```
```{r eval=FALSE, include=FALSE}
str(field_gcc)
str(field_gcc_ndvi)
```


```{r}
field_gcc_long$burn_year = as.factor(field_gcc_long$burn_year)
field_ndvi_long$burn_year= as.factor(field_ndvi_long$burn_year)
field_gcc_long$buf_res = as.factor(field_gcc_long$buf_res)
field_ndvi_long$buf_res = as.factor(field_ndvi_long$buf_res)
```



# Define graph Theme and color Palette 

```{r}

reg_theme = theme_bw() + theme(legend.position = "none") +
  theme(plot.margin = unit(c(t = 0.3, r = 0.3, b = 0.3, l = 0.1), "cm")) +
  theme(axis.title.x = element_text(size = 10, hjust = 0.5, vjust = -0.2),
        axis.title.y = element_text(size = 10, hjust = 0.5, vjust = 1.1),
        axis.text.x = element_text(size = 9, color = "black"),
        axis.text.y = element_text(size = 9, color = "black"),
        panel.background = element_blank(), 
        axis.line = element_line(colour = "black"))

```

```{r}
cbbPalette <- c("#000000", "#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#D55E00", "#CC79A7")
```


# Analysis: Biomass ~ GCC

```{r}
head(field_gcc_long1)
```
## Model GCC
```{r}
fit1 = lm('biomass_log ~ gcc + res_meters + gcc:res_meters', data = field_gcc_long1)
```

### Summary

```{r}
summary(fit1)
```
```{r}
summary(fit1)$coef
```

```{r}
summary(fit1)$r.squared 
```


####### Calculated y-intercept and slope
```{r}
int_25cm = coef(fit1)[1]
int_50cm = coef(fit1)[1] + coef(fit1)[3]
int_1m = coef(fit1)[1] + coef(fit1)[4]
int_3m = coef(fit1)[1] + coef(fit1)[5]
int_5m = coef(fit1)[1] + coef(fit1)[6]
int_10m = coef(fit1)[1] + coef(fit1)[7]

int_25cm
int_50cm
int_1m
int_3m
int_5m
int_10m

slope_25cm = coef(fit1)[2]
slope_50cm = coef(fit1)[2] + coef(fit1)[8]
slope_1m = coef(fit1)[2] + coef(fit1)[9]
slope_3m = coef(fit1)[2] + coef(fit1)[10]
slope_5m = coef(fit1)[2] + coef(fit1)[11]
slope_10m = coef(fit1)[2] + coef(fit1)[12]
slope_25cm
slope_50cm
slope_1m
slope_3m
slope_5m
slope_10m

```

```{r}
coef(fit1)
```
```{r}
summary(fit1)$coefficients[,4]  
```


### Pulling out r-squared for each group
reference: https://drsimonj.svbtle.com/running-a-model-on-separate-groups


```{r}
field_gcc_long1 %>% 
  nest(-buf_res) %>% 
  mutate(fit = map(data, ~ lm(biomass_log ~ gcc, data = .)),
         results = map(fit, glance)) %>% 
  unnest(results)
```

```{r}
field_gcc_long1 %>% 
  nest(-buf_res) %>% 
  mutate(fit = map(data, ~ lm(biomass_log ~ gcc, data = .)),
         results = map(fit, glance)) %>% 
  unnest(results) %>% 
  ggplot(aes(x = factor(buf_res), y = r.squared)) +
    geom_bar(stat = "identity") +
    labs(x = "Buffer Resolution", y = expression(R^{2}))
```

#### Residuals
```{r}
# Get the fitted value for each observation
field_gcc_long1$fit1 = fitted(fit1)
# Get the residuals of the model
field_gcc_long1$res1 = residuals(fit1)
```

```{r}
qplot(x = fit1, y = res1, data = field_gcc_long1,
main = "Residuals vs Fitted Values")
```

```{r}
qplot(x = gcc, y = res1, data = field_gcc_long1,
xlab = "gcc",
ylab = "Residuals",
main = "Residuals vsgcc")
```
```{r}
qplot(x = factor(1), y = res1, data = field_gcc_long1, geom = "boxplot")
```

```{r}
qqnorm(field_gcc_long1$res1, main = "Normal QQ Plot of Residuals")
qqline(field_gcc_long1$res1) # add reference line to the qq plot
```

```{r}
plot(fit1, which = 1) # residual vs fitted values
```

```{r}
plot(fit1, which = 2) # qqnorm plot of residuals
```
### Plot

```{r fig.height=5, fig.width=6}
plot_fit1 = ggplot(data = field_gcc_long1, aes(x = gcc, y = biomass_log, color = res_meters)) + 
  geom_smooth(method="lm", se=TRUE, alpha=0.1) +
  geom_point(size = 1) +
  scale_color_manual(values=cbbPalette) +
    labs(y=expression(atop("Log-transformed C biomass", paste((g~C~m^-2)))), x= "GCC", color = "Buffer resolution") +
  theme_bw() + 
  theme(legend.position = "bottom") +
  guides(colour = guide_legend(nrow = 1))

plot_fit1
  
```

```{r}
plot_fit1.1 = ggplot(data = field_gcc_long1, aes(x = gcc, y = biomass_log, color = res_meters)) + 
  geom_smooth(method="lm", se=TRUE, alpha=0.1) +
  geom_point(size = 1) +
  scale_color_manual(values=cbbPalette) +
    labs(y=expression(atop("Log-transformed C biomass", paste((g~C~m^-2)))), x= "GCC", color = "Buffer resolution") +
  reg_theme

plot_fit1.1
  
```

```{r eval=FALSE, include=FALSE}
ggsave("../figures/2020-08-19_gcc-biomass-resolution.jpeg", plot = plot_fit1, width = 6, height = 5, units = c("in"), dpi=600 )
```

# Analysis: Biomass ~ NDVI


```{r}
head(field_ndvi_long1)
```

## Model NDVI
```{r}
fit2 = lm(biomass_log ~ ndvi + res_meters + ndvi:res_meters, data = field_ndvi_long1)
```

### Summary
```{r}
summary(fit2)
```
```{r}
summary(fit2)$r.squared 
```
```{r}
int_25cm = coef(fit2)[1]
int_50cm = coef(fit2)[1] + coef(fit2)[3]
int_1m = coef(fit2)[1] + coef(fit2)[4]
int_3m = coef(fit2)[1] + coef(fit2)[5]
int_5m = coef(fit2)[1] + coef(fit2)[6]
int_10m = coef(fit2)[1] + coef(fit2)[7]

int_25cm
int_50cm
int_1m
int_3m
int_5m
int_10m

slope_25cm = coef(fit2)[2]
slope_50cm = coef(fit2)[2] + coef(fit2)[8]
slope_1m = coef(fit2)[2] + coef(fit2)[9]
slope_3m = coef(fit2)[2] + coef(fit2)[10]
slope_5m = coef(fit2)[2] + coef(fit2)[11]
slope_10m = coef(fit2)[2] + coef(fit2)[12]
slope_25cm
slope_50cm
slope_1m
slope_3m
slope_5m
slope_10m

```


### Pulling out r-squared for each group
reeference: https://drsimonj.svbtle.com/running-a-model-on-separate-groups

lm(formula = biomass_log ~ ndvi + res_meters + ndvi:res_meters, 
    data = field_ndvi_long1
```{r}
 field_ndvi_long1 %>% 
  nest(-buf_res) %>% 
  mutate(fit = map(data, ~ lm(biomass_log ~ ndvi, data = .)),
         results = map(fit, glance)) %>% 
  unnest(results)
```

```{r}
 field_ndvi_long1 %>% 
  nest(-buf_res) %>% 
  mutate(fit = map(data, ~ lm(biomass_log ~ ndvi, data = .)),
         results = map(fit, glance)) %>% 
  unnest(results) %>% 
  ggplot(aes(x = factor(buf_res), y = r.squared)) +
    geom_bar(stat = "identity") +
    labs(x = "Buffer Resolution", y = expression(R^{2}))
```


#### Residuals
```{r}
# Get the fitted value for each observation
field_ndvi_long1$fit2 = fitted(fit2)
# Get the residuals of the model
field_ndvi_long1$res2 = residuals(fit2)
```

```{r}
qplot(x = fit2, y = res2, data = field_ndvi_long1,
main = "Residuals vs Fitted Values")
```

```{r}
qplot(x = ndvi, y = res2, data = field_ndvi_long1,
xlab = "ndvi",
ylab = "Residuals",
main = "Residuals vs ndvi")
```
```{r}
qplot(x = factor(1), y = res2, data = field_ndvi_long1, geom = "boxplot")
```

```{r}
qqnorm(field_ndvi_long1$res2, main = "Normal QQ Plot of Residuals")
qqline(field_ndvi_long1$res2) # add reference line to the qq plot
```

```{r}
plot(fit2, which = 1) # residual vs fitted values
```

```{r}
plot(fit2, which = 2) # qqnorm plot of residuals
```

### Plot

```{r fig.height=5, fig.width=6}
plot_fit2 = ggplot(data = field_ndvi_long1, aes(x = ndvi, y = biomass_log, color = res_meters)) + 
  geom_smooth(method="lm", se=TRUE, alpha=0.1) +
  geom_point(size = 1) +
  scale_color_manual(values=cbbPalette) +
    labs(y=expression(atop("Log-transformed C biomass", paste((g~C~m^-2)))), x= "NDVI", color = "Buffer resolution") +
  theme_bw() + 
  theme(legend.position = "bottom") +
  guides(colour = guide_legend(nrow = 1))

plot_fit2
  
```

```{r}
plot_fit2.1 = ggplot(data = field_ndvi_long1, aes(x = ndvi, y = biomass_log, color = res_meters)) + 
  geom_smooth(method="lm", se=TRUE, alpha=0.1) +
  geom_point(size = 1) +
  scale_color_manual(values=cbbPalette) +
    labs(y=expression(atop("Log-transformed C biomass", paste((g~C~m^-2)))), x= "NDVI", color = "Buffer resolution") +
  reg_theme

plot_fit2.1
  
```

```{r eval=FALSE, include=FALSE}
ggsave("../figures/2020-08-19_ndvi-biomass-resolution.jpeg", plot = plot_fit2, width = 6, height = 5, units = c("in"), dpi=600 )
```


# Manuscript Figures 
```{r}
legend_res <- get_legend(plot_fit1) 
```

## GCC and NDVI

```{r fig.height=7, fig.width=6}

fig1 = cowplot::plot_grid(plot_fit1.1, plot_fit2.1, labels = "AUTO", align = "hv", label_size = 12, ncol= 1)  
  
fig1
```

```{r fig.height=7, fig.width=6}
fig1.1 = cowplot::plot_grid(fig1, legend_res, ncol=1, rel_heights = c(1, 0.04))  

fig1.1
```

```{r eval=FALSE, include=FALSE}
ggsave("../figures/2020-08-20_Biomass-ndvi-gcc.jpeg", plot = fig1.1, width = 6, height = 7, units = c("in"), dpi=600 )
```
