==> scripts/3.4_Field-UAV-BasalArea.Rmd <==
---
title: "Field and UAv Data Analysis"
author: "Anna Talucci"
date: "2/28/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```
# Overview

Analysis of field data (shrub basal area, and tree basal area)

# Pakcages
```{r}
library(tidyr)
library(dplyr)
library(ggplot2)
library(stringr)
library(lme4)
library(cowplot)
library(lme4)
library(lsmeans)
library(DHARMa)
```


# Functions

```{r}
numextract <- function(string){ 
  str_extract(string, "\\-*\\d+\\.*\\d*")
} 

```
# Data

## Field Data
```{r}
field_data = read.csv("../data/field_subset/field_data_all.csv")
```

```{r}
head(field_data)
```

## UAV Data

```{r}
gcc_all = read.csv("../data/extracted_data/gcc_all.csv", strip.white = TRUE)
ndvi_all = read.csv("../data/extracted_data/2020-08-10_ndvi_all_pix4D.csv", strip.white = TRUE)
```

```{r}
head(gcc_all)
head(ndvi_all)
```
### Pull out GCC and NDVI by buffer size

```{r}
gcc_long = gcc_all %>% 
  dplyr::select(ID:mean_gcc_10m) %>%
  pivot_longer(-ID, names_to = "buffer", values_to = "gcc") %>%
  mutate(buf_res = numextract(buffer))

gcc_long
```

```{r}
ndvi_long = ndvi_all %>% 
  dplyr::select(ID:mean_ndvi_10m) %>%
  pivot_longer(-ID, names_to = "buffer", values_to = "ndvi") %>%
  mutate(buf_res = numextract(buffer))

ndvi_long
```

### Combine Field and UAV data

```{r}
gcc_ndvi = merge(gcc_all, ndvi_all, by="ID")
gcc_ndvi
field_gcc = merge(field_data, gcc_all, by="ID") 
field_gcc
field_gcc_ndvi = merge(field_data, gcc_ndvi,  "ID")
field_gcc_ndvi
```
### Field data +  long data
```{r}
field_gcc_long = merge(field_data, gcc_long, by="ID")

field_gcc_long1 = field_gcc_long %>% 
  dplyr::select(ID:burn_year, live_woody_biomass_C_g_m, buffer:buf_res) %>%
  mutate(res_meters = ifelse(buf_res %in% 25, "25cm",
                                    ifelse(buf_res %in% 50, "50cm",
                                    ifelse(buf_res %in% 1, "1m",
                                    ifelse(buf_res %in% 3, "3m",
                                    ifelse(buf_res %in% 5, "5m", "10m")))))) %>%
  mutate(res_meters = factor(res_meters, levels = c("25cm", "50cm", "1m", "3m", "5m", "10m"))) %>%
  mutate(biomass_log = log(live_woody_biomass_C_g_m))

head(field_gcc_long1)
```

```{r}
field_ndvi_long = merge(field_data, ndvi_long, by="ID")

field_ndvi_long1 = field_ndvi_long %>% 
  dplyr::select(ID:burn_year, live_woody_biomass_C_g_m, buffer:buf_res) %>%
  mutate(res_meters = ifelse(buf_res %in% 25, "25cm",
                                    ifelse(buf_res %in% 50, "50cm",
                                    ifelse(buf_res %in% 1, "1m",
                                    ifelse(buf_res %in% 3, "3m",
                                    ifelse(buf_res %in% 5, "5m", "10m")))))) %>%
  mutate(res_meters = factor(res_meters, levels = c("25cm", "50cm", "1m", "3m", "5m", "10m"))) %>%
  mutate(biomass_log = log(live_woody_biomass_C_g_m))

head(field_ndvi_long1)
```

### Variables as factor

```{r}
field_gcc$burn_year = as.factor(field_gcc$burn_year)
field_gcc_ndvi$burn_year = as.factor(field_gcc_ndvi$burn_year)

```

```{r}
field_gcc_long$burn_year = as.factor(field_gcc_long$burn_year)
field_ndvi_long$burn_year= as.factor(field_ndvi_long$burn_year)
field_gcc_long$buf_res = as.factor(field_gcc_long$buf_res)
field_ndvi_long$buf_res = as.factor(field_ndvi_long$buf_res)
```

# Subset data
```{r}
head(field_gcc)
```

```{r}
field_gcc1 = field_gcc %>% 
  dplyr::select(ID:burn_year, live_tree_ba_ha, shrub_ba_ha, mean_gcc_10m) %>%
  mutate(tree_ba_ha_log = log(live_tree_ba_ha)) %>%
  mutate(shrub_ba_ha_log = log(shrub_ba_ha))

head(field_gcc1)
```
```{r}
field_ndvi1 = field_gcc_ndvi %>% 
  dplyr::select(ID:burn_year, live_tree_ba_ha, shrub_ba_ha, mean_ndvi_10m) %>%
  mutate(tree_ba_ha_log = log(live_tree_ba_ha)) %>%
  mutate(shrub_ba_ha_log = log(shrub_ba_ha))

field_ndvi1
```

# Define graph Theme and color Palette 

```{r}
comp_theme = theme_bw() + theme(legend.position = "none") +  
  theme(panel.grid.major = element_blank()) +
    theme(axis.title.y = element_text(size = 11, hjust = 0.5, vjust = 1.1),
        axis.text.x = element_text(size = 10, color = "black"),
        axis.text.y = element_text(size = 10, color = "black"),
        axis.line = element_line(colour = "black"))
```
 
```{r}

reg_theme = theme_bw() + theme(legend.position = "none") +
  theme(plot.margin = unit(c(t = 0.3, r = 0.3, b = 0.3, l = 0.1), "cm")) +
  theme(axis.title.x = element_text(size = 11, hjust = 0.5, vjust = -0.1),
        axis.title.y = element_text(size = 11, hjust = 0.5, vjust = 1.1),
        axis.text.x = element_text(size = 10, color = "black"),
        axis.text.y = element_text(size = 10, color = "black"),
        panel.background = element_blank(), 
        axis.line = element_line(colour = "black"))

```

```{r}
cbbPalette <- c("#000000", "#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#D55E00", "#CC79A7")
```

# Analysis: Vegetation cover

```{r}
head(field_gcc1)
head(field_ndvi1)
```


## GCC Models


### 1. Model: tree  BA

```{r}
fit1 = lm(live_tree_ba_ha ~ mean_gcc_10m, data = field_gcc1)
```

#### Residuals
```{r}
# Get the fitted value for each observation
field_gcc1$fit1 = fitted(fit1)
# Get the residuals of the model
field_gcc1$res1 = residuals(fit1)
```

```{r}
qplot(x = fit1, y = res1, data = field_gcc1,
main = "Residuals vs Fitted Values")
```

```{r}
qplot(x = mean_gcc_10m, y = res1, data = field_gcc1,
xlab = "gcc",
ylab = "Residuals",
main = "Residuals vsgcc")
```
```{r}
qplot(x = factor(1), y = res1, data = field_gcc1, geom = "boxplot")
```

```{r}
qqnorm(field_gcc1$res1, main = "Normal QQ Plot of Residuals")
qqline(field_gcc1$res1) # add reference line to the qq plot
```

```{r}
plot(fit1, which = 1) # residual vs fitted values
```

```{r}
plot(fit1, which = 2) # qqnorm plot of residuals
```



####  summary
```{r}
summary(fit1)
```


```{r}
fit1_df = data.frame(summary(fit1)$coefficients[,], confint(fit1), extractAIC(fit1))
fit1_df
fit1_df2 = tibble::rownames_to_column(fit1_df, "coefficent")
fit1_df3 = fit1_df2 %>% 
  mutate(model = "fit1_tree_ba") %>% 
  mutate(response = "GCC" ) %>%  
  mutate(rmse = RMSE(fit1)) %>%
  rename(estimate = Estimate, se = Std..Error, t_value = t.value,  p_value = Pr...t.., lcl = X2.5.., ucl = X97.5.., AIC = extractAIC.fit1.) %>% dplyr::select(model:response, coefficent:rmse)
fit1_df3
```
```{r}
summary(fit1)$r.squared 
```
#### Graph lm

```{r}
p1 = ggplot(data = field_gcc1, aes(x = mean_gcc_10m, y = live_tree_ba_ha)) + 
    geom_point(size = 2) +  
  stat_smooth(method = "lm", col = "black", alpha = .25) +
    labs(y=expression(atop("Tree basal area", paste((m^2/ha)))), x= "GCC") +
    reg_theme +  guides(color=guide_legend(override.aes=list(fill=NA))) 

p1
  
```
(expression(paste(Delta~AIC, "=32.5"))

### 2. Model: shrub  BA
```{r}
fit2 = lm(shrub_ba_ha ~ mean_gcc_10m , data = field_gcc1)
```

#### Residuals
```{r}
# Get the fitted value for each observation
field_gcc1$fit2 = fitted(fit2)
# Get the residuals of the model
field_gcc1$res2 = residuals(fit2)
```

```{r}
qplot(x = fit2, y = res2, data = field_gcc1,
main = "Residuals vs Fitted Values")
```

```{r}
qplot(x = mean_gcc_10m, y = res2, data = field_gcc1,
xlab = "gcc",
ylab = "Residuals",
main = "Residuals vs  gcc")
```
```{r}
qplot(x = factor(1), y = res2, data = field_gcc1, geom = "boxplot")
```

```{r}
qqnorm(field_gcc1$res2, main = "Normal QQ Plot of Residuals")
qqline(field_gcc1$res2) # add reference line to the qq plot
```

```{r}
plot(fit2, which = 1) # residual vs fitted values
```

```{r}
plot(fit2, which = 2) # qqnorm plot of residuals
```
#### Summary
```{r}
summary(fit2)
```


```{r}
fit2_df = data.frame(summary(fit2)$coefficients[,], confint(fit2), extractAIC(fit2))
fit2_df
fit2_df2 = tibble::rownames_to_column(fit2_df, "coefficent")
fit2_df3 = fit2_df2 %>% 
  mutate(model = "fit2_shrub_ba") %>% 
  mutate(response = "GCC" ) %>%  
  mutate(rmse = RMSE(fit2)) %>%
  rename(estimate = Estimate, se = Std..Error, t_value = t.value,  p_value = Pr...t.., lcl = X2.5.., ucl = X97.5.., AIC = extractAIC.fit2.) %>% dplyr::select(model:response, coefficent:rmse)
fit2_df3
```
```{r}
summary(fit2)$r.squared 
```
#### Graph lm
```{r}
p2 = ggplot(data = field_gcc1, aes(x = mean_gcc_10m, y = shrub_ba_ha)) + 
    geom_point(size = 2) +  
  stat_smooth(method = "lm", col = "black", alpha = .25) +
    labs(y=expression(atop("Shrub basal area", paste((m^2/ha)))), x= "GCC") +
    reg_theme +  guides(color=guide_legend(override.aes=list(fill=NA))) 

p2
  
```

## NDVI Models

### 3. Model: tree BA

```{r}
fit3 = lm(live_tree_ba_ha ~ mean_ndvi_10m, data = field_ndvi1)
```

#### Residuals
```{r}
# Get the fitted value for each observation
field_ndvi1$fit3 = fitted(fit3)
# Get the residuals of the model
field_ndvi1$res3 = residuals(fit3)
```

```{r}
qplot(x = fit3, y = res3, data = field_ndvi1,
main = "Residuals vs Fitted Values")
```

```{r}
qplot(x = mean_ndvi_10m, y = res3, data = field_ndvi1,
xlab = "ndvi",
ylab = "Residuals",
main = "Residuals vsgcc")
```
```{r}
qplot(x = factor(1), y = res3, data = field_ndvi1, geom = "boxplot")
```

```{r}
qqnorm(field_ndvi1$res3, main = "Normal QQ Plot of Residuals")
qqline(field_ndvi1$res3) # add reference line to the qq plot
```

```{r}
plot(fit3, which = 1) # residual vs fitted values
```

```{r}
plot(fit3, which = 2) # qqnorm plot of residuals
```

#### Summary
```{r}
summary(fit3)
```
```{r}
summary(fit3)$r.squared 
```

```{r}
fit3_df = data.frame(summary(fit3)$coefficients[,], confint(fit3), extractAIC(fit3))
fit3_df
fit3_df2 = tibble::rownames_to_column(fit3_df, "coefficent")
fit3_df3 = fit3_df2 %>% 
  mutate(model = "fit3_tree_ba") %>% 
  mutate(response = "NDVI" ) %>%  
  mutate(rmse = RMSE(fit3)) %>%
  rename(estimate = Estimate, se = Std..Error, t_value = t.value,  p_value = Pr...t.., lcl = X2.5.., ucl = X97.5.., AIC = extractAIC.fit3.) %>% dplyr::select(model:response, coefficent:rmse)
fit3_df3
```


#### Graph lm
```{r}
p3 = ggplot(data = field_ndvi1, aes(x = mean_ndvi_10m, y = live_tree_ba_ha)) + 
    geom_point(size = 2) +  
  stat_smooth(method = "lm", col = "black", alpha = .25) +
    labs(y=expression(atop("Tree basal area", paste((m^2/ha)))), x= "NDVI") +
    reg_theme +  guides(color=guide_legend(override.aes=list(fill=NA))) 

p3
  
```



### 4. Model: shrub BA



```{r}
fit4 = lm(shrub_ba_ha ~ mean_ndvi_10m, data = field_ndvi1)
```

#### Residuals
```{r}
# Get the fitted value for each observation
field_ndvi1$fit4 = fitted(fit4)
# Get the residuals of the model
field_ndvi1$res4 = residuals(fit4)
```

```{r}
qplot(x = fit4, y = res4, data = field_ndvi1,
main = "Residuals vs Fitted Values")
```

```{r}
qplot(x = mean_ndvi_10m, y = res4, data = field_ndvi1,
xlab = "ndvi",
ylab = "Residuals",
main = "Residuals vs  gcc")
```
```{r}
qplot(x = factor(1), y = res4, data = field_ndvi1, geom = "boxplot")
```

```{r}
qqnorm(field_ndvi1$res4, main = "Normal QQ Plot of Residuals")
qqline(field_ndvi1$res4) # add reference line to the qq plot
```

```{r}
plot(fit4, which = 1) # residual vs fitted values
```

```{r}
plot(fit4, which = 2) # qqnorm plot of residuals
```

#### Summary

```{r}
summary(fit4)
```

```{r}
summary(fit4)$r.squared 
```

```{r}
fit4_df = data.frame(summary(fit4)$coefficients[,], confint(fit4), extractAIC(fit4))
fit4_df
fit4_df2 = tibble::rownames_to_column(fit4_df, "coefficent")
fit4_df3 = fit4_df2 %>% 
  mutate(model = "fit4_shrub_ba") %>% 
  mutate(response = "NDVI" ) %>%  
  mutate(rmse = RMSE(fit4)) %>%
  rename(estimate = Estimate, se = Std..Error, t_value = t.value,  p_value = Pr...t.., lcl = X2.5.., ucl = X97.5.., AIC = extractAIC.fit4.) %>% dplyr::select(model:response, coefficent:rmse)
fit4_df3
```

#### Graph lm

```{r}
p4 = ggplot(data = field_ndvi1, aes(x = mean_ndvi_10m, y = shrub_ba_ha)) + 
    geom_point(size = 2) +  
  stat_smooth(method = "lm", col = "black", alpha = .25) +
    labs(y=expression(atop("Shrub basal area", paste((m^2/ha)))), x= "NDVI") +
    reg_theme +  guides(color=guide_legend(override.aes=list(fill=NA))) 

p4
  
```

# Combine Summary output lm

```{r}
fit_sum1 = rbind(fit1_df3, fit2_df3)
fit_sum2 = rbind(fit3_df3, fit4_df3)
fit_sum3 = rbind(fit_sum1, fit_sum2)

fit_sum3
```

```{r eval=FALSE, include=FALSE}
write.csv(fit_sum3, '../outputs/model_summary/2020-08-21_vegetation-cover-gcc-ndvi_lm.csv', row.names = F)

```


# Manuscript Figures

## lm Figure
```{r fig.height=5, fig.width=6}

lm_grid_veg1 = cowplot::plot_grid(p1, p2, p3, p4, labels = c("A", "B", "C", "D"), align = "hv", ncol= 2) 
lm_grid_veg1

```


```{r eval=FALSE, include=FALSE}
ggsave("../figures/2020-08-21_BA-gcc-ndvi_lm.jpeg", plot = lm_grid_veg1, width = 6, height = 5, units = c("in"), dpi=600 )
```
